### 2021-04-09 #3159 @whedon
[REVIEW]: fctSnPM: Factorial ANOVA and post-hoc tests for Statistical nonParametric Mapping in MATLAB
**Submitting author:** @tramarobin (<a href="http://orcid.org/0000-0001-6807-0505">Robin Trama</a>)
**Repository:** <a href="https://github.com/tramarobin/fctSnPM" target ="_blank">https://github.com/tramarobin/fctSnPM</a>
**Version:** v2.0
