+++ README.md
@@ -1,4 +1,4 @@
-# fctSPM [![View fctSPM on File Exchange](https://www.mathworks.com/matlabcentral/images/matlab-file-exchange.svg)](https://www.mathworks.com/matlabcentral/fileexchange/77945-fctspm)
+# fctSPM [![View fctSPM on File Exchange](https://www.mathworks.com/matlabcentral/images/matlab-file-exchange.svg)](https://www.mathworks.com/matlabcentral/fileexchange/77945-fctspm)[![status](https://joss.theoj.org/papers/ea923a728497b806dbb59a0c4c0b76cc/status.svg)](https://joss.theoj.org/papers/ea923a728497b806dbb59a0c4c0b76cc)
 Using spm1d package (v.0.4.3), computes ANOVA and post-hoc tests from anova1 to anova3rm, with a non-parametric approach (permutation tests).
 The type of anova (if required) and post-hoc are choosen regarding the independant or repeated measure effects given in parameters.
 The function automatically adapts to 1D and 2D data.
@@ -41,7 +41,7 @@ This function synthetises the main and interaction effects to display only the s
 For post-hoc for interaction effects, the main effect is also displayed if located elsewhere than the interaction effect.
 
 ## Citing fctSPM ##
-- This function : in progress   
+- This function : under review in JOSS   
 - for spm1d : Pataky TC (2010). Generalized n-dimensional biomechanical field analysis using statistical parametric mapping. Journal of Biomechanics 43, 1976-1982.   
 - for permutation tests : Nichols TE, Holmes AP (2002). Nonparametric permutation tests for functional neuroimaging: a primer with examples. Human Brain Mapping 15(1), 1–25.   
 - spm1d package for matlab can be downloaded at : https://github.com/0todd0000/spm1dmatlab   
