### 2021-04-10 #3159 @0todd0000
@cMadan: I have just read the journal's COI policy. I am writing to disclose a potential COI.

The in-review repository [fctSPM](https://github.com/tramarobin/fctSPM) is an add-on for one of my own packages: [spm1dmatlab](https://github.com/0todd0000/spm1dmatlab), and in related [public documentation](https://spm1d.org/Addons.html) I have linked to the in-review [fctSPM](https://github.com/tramarobin/fctSPM) package. Last, I have had brief email correspondence with @tramarobin regarding [fctSPM](https://github.com/tramarobin/fctSPM) and how it could be integrated with [spm1dmatlab](https://github.com/0todd0000/spm1dmatlab). The result of that conversation was the suggestion that [fctSPM](https://github.com/tramarobin/fctSPM) be developed and maintained independently of [spm1dmatlab](https://github.com/0todd0000/spm1dmatlab), and that I would link to [fctSPM](https://github.com/tramarobin/fctSPM) through online documentation, as noted above. At the moment I can't seem to find a record of this correspondence, but could provide this record if necessary.

In my opinion, this does not represent a COI because (1) I have worked neither directly nor indirectly with @tramarobin on [fctSPM](https://github.com/tramarobin/fctSPM), and (2) it would be in [spm1dmatlab](https://github.com/0todd0000/spm1dmatlab)'s interest to have only rigorously reviewed add-ons.

Regardless, I am now disclosing this potential COI for your consideration. Please advise.
