+++ Examples/D2_ANOVA1rm.m
@@ -4,9 +4,8 @@
 clear
 close all
 clc
-
-% path
-addAbovePath
+% add source code path
+addpath(genpath("../src"))
 
 % data
 load ExampleDatas
