+++ Examples/D1_ANOVA3_2rm.m
@@ -4,8 +4,8 @@ clear
 close all
 clc
 
-% path
-addAbovePath
+% add source code path
+addpath(genpath("../src"))
 
 % data
 load ExampleDatas
