+++ Examples/D2_ANOVA1.m
@@ -5,8 +5,9 @@ clear
 close all
 clc
 
-% path
-addAbovePath
+% add source code path
+addpath(genpath("../src"))
+
 
 % data
 load ExampleDatas
