+++ README.md
@@ -14,6 +14,7 @@ fctSPM(data, independantEffects, repeatedMeasuresEffects, varargin)
 - [Why fctSPM](#Why-fctSPM)
 - [Citing fctSPM](#Citing-fctSPM)
 - [MATLAB Release Compatibility](#MATLAB-Release-Compatibility)
+- [Outputs](#Outputs)
 - [Examples](#Examples)
 - [Using fctSPM](#Using-fctSPM)
 - [Obligatory INPUTS](#Obligatory-INPUTS)
@@ -49,6 +50,89 @@ For post-hoc for interaction effects, the main effect is also displayed if locat
 
 ## MATLAB Release Compatibility ##
 Compatible from Matlab R2017b
+
+
+## Outputs ##
+This fonction creates two folders:
+* `ANOVA` named after the type of ANOVA performed on the data (ANOVA1rm if the analysis was a one-way ANOVA with repeated measures, no ANOVA folder is created if the analysis is a t-test).  
+* `Post hoc` for the post-hoc analysis.  
+
+Each folder is composed of two different types of files : 
+* `.m` files in which all data and statistical analysis can be found
+* `.TIF` and `.fig` files representing the statistical analysis on different graphics.  
+
+### ANOVA folder ###
+`anova.m` file is composed of different fields (same for 1D and 2D) :  
+* `type is the type of ANOVA performed
+* `effectNames` is a structure (one cell for each effect) that represent the names of the effects tested (mains and interactions) 
+* `alpha` is the alpha risk used for the anova
+* `alphaOriginal` is the alpha risk choosen for the anova (default is 0.05 (5%)). Warning message is displayed if this value is modified.
+* `nIterations` is the number of iterations performed for the anova.
+* `maxIterations` is the number of maximal iterations possible for the anova.
+* `Fcontinuum` is a structure that represent the F-value for each node
+* `Fthreshold` is a structure that represent the statistical threshold for the F-values (statistical inference)
+* `Fsignificant` is a structure that contains the logical for the significance (Fcontinuum > Fthreshold) of each effect of the ANOVA (1 if significant, 0 if not).  
+* `clusterLocation` is a structure (one for each significant cluster) that contains the location (start and end as indexes) of each significant cluster.
+* `clusterP` is a structure (one for each significant cluster) that contains the p-value of each significant cluster.
+`clusterLocation` and `clusterP` are interpretable only for 1D data. Use figures to interpret 2D data significance.
+
+Each effect is also display on a specific figure in `.TIF` format named after `effectNames`, a floder named FIG is also created and contains the figures in `.fig` format.   
+For 1D:  The curve represents the `Fcontinuum`, the horizontal line is the `Fthreshold`, the highlighted curve(s) in blue represent(s) the significant cluster(s), and the vertical lines are the start and end indexes.
+For 2D: The map represents the `Fcontinuum`, with a colorbar thresholded at `Fthreshold`, the white clusters represent the significant clusters.
+
+
+### Post hoc folder ###
+This folder contains additional folders (0 (for anova1), 3 (for anova2) or 7 (for anova3)) that contain figures and metrics representing the different post-hoc tests.
+`posthoc.m` file is composed of different fields :
+* `data.names` is a structure that contains the name of the conditions used in the analysis (\cap is the union of different conditions for interactions).
+* `data.continuum` is a structure that contains the data used in the analysis.
+
+For the following outputs, the number of cells of the structure correspond to the number of t-tests performed, one t-test corresponds to one cell.
+* `differences.names` is the name of the conditions (the first minus the second) used in the differences and t-tests.
+* `differences.continuum` is the data used to plot differences.
+* `differences.continuumRelative` is the data used to plot relative differences
+* `differences.ES` is the effect size that correspond to the differences (Hedge's g)
[…]
