+++ README.md
@@ -63,7 +63,7 @@ Each folder is composed of two different types of files :
 
 ### ANOVA folder ###
 `anova.m` file is composed of different fields (same for 1D and 2D) :  
-* `type is the type of ANOVA performed
+* `type` is the type of ANOVA performed
 * `effectNames` is a structure (one cell for each effect) that represent the names of the effects tested (mains and interactions) 
 * `alpha` is the alpha risk used for the anova
 * `alphaOriginal` is the alpha risk choosen for the anova (default is 0.05 (5%)). Warning message is displayed if this value is modified.
@@ -73,16 +73,17 @@ Each folder is composed of two different types of files :
 * `Fthreshold` is a structure that represent the statistical threshold for the F-values (statistical inference)
 * `Fsignificant` is a structure that contains the logical for the significance (Fcontinuum > Fthreshold) of each effect of the ANOVA (1 if significant, 0 if not).  
 * `clusterLocation` is a structure (one for each significant cluster) that contains the location (start and end as indexes) of each significant cluster.
-* `clusterP` is a structure (one for each significant cluster) that contains the p-value of each significant cluster.
+* `clusterP` is a structure (one for each significant cluster) that contains the p-value of each significant cluster.  
+
 `clusterLocation` and `clusterP` are interpretable only for 1D data. Use figures to interpret 2D data significance.
 
 Each effect is also display on a specific figure in `.TIF` format named after `effectNames`, a floder named FIG is also created and contains the figures in `.fig` format.   
-For 1D:  The curve represents the `Fcontinuum`, the horizontal line is the `Fthreshold`, the highlighted curve(s) in blue represent(s) the significant cluster(s), and the vertical lines are the start and end indexes.
-For 2D: The map represents the `Fcontinuum`, with a colorbar thresholded at `Fthreshold`, the white clusters represent the significant clusters.
+For 1D:  The curve represents the `Fcontinuum`, the horizontal line is the `Fthreshold`, the highlighted parts of the curve in blue represents the significant cluster(s), and the vertical lines are the start and end indexes.
+For 2D: The map represents the `Fcontinuum`, with a colorbar maximum at `Fthreshold`, the white clusters represent the significant clusters.
 
 
 ### Post hoc folder ###
-This folder contains additional folders (0 (for anova1), 3 (for anova2) or 7 (for anova3)) that contain figures and metrics representing the different post-hoc tests.
+This folder contains additional folders (0 (for anova1), 3 (for anova2) or 7 (for anova3)) that contain figures and metrics representing the different post-hoc tests.  
 `posthoc.m` file is composed of different fields :
 * `data.names` is a structure that contains the name of the conditions used in the analysis (\cap is the union of different conditions for interactions).
 * `data.continuum` is a structure that contains the data used in the analysis.
@@ -98,20 +99,21 @@ For the following outputs, the number of cells of the structure correspond to th
 * `tTests.nWarning` represents the number of warnings displyed during the analysis : 0 is OK, 1 means the number of iterations was reduced but the original alpha risk was conserved, 2 means that the number of iterations was reduced and the original alpha was reduced. In this case, more subjects are required to performed the analysis. 
 * `tTests.alphaOriginal` is the alpha risk choosen for the post hoc tests (default is 0.05 (5%)/ number of comparisons). Warning message is displayed if this value is modified.
 * `tTests.alpha` is the alpha risk used for the post hoc tests
-* `tTests.alphaOriginal` is the alpha risk choosen for the anova (default is 0.05 (5%)). Warning message is displayed if this value is modified.
 * `tTests.nIterations` is the number of iterations performed for the t-test.
 * `tTests.maxIterations` is the number of maximal iterations possible for the t-test.
[…]
