### 2021-04-13 1d72218
modifications for review
readme and fctSPM.m : add the outputs description

outputs were modified to be easier to interpret

Examples : default savedir was modified. When non specified, dialog box ask where to save the files.
