+++ src/chooseAnova.m
@@ -1,4 +1,4 @@
-function [ANOVA,namesEffect,testANOVA]=chooseAnova(maps1d,sujets,nEffects,nRm,indicesEffects,eNames)
+function [ANOVA,anova]=chooseAnova(maps1d,sujets,nEffects,nRm,indicesEffects,eNames)
 
 
 if nEffects==1
@@ -6,17 +6,17 @@ if nEffects==1
         indicesAnovaEffects{1}=1;
         if nRm==0
             ANOVA=spm1d.stats.nonparam.anova1(maps1d,indicesEffects(:,1));
-            testANOVA.name='ANOVA1';
+            anova.type='ANOVA1';
             disp('ANOVA1')
         elseif nRm==1
             ANOVA=spm1d.stats.nonparam.anova1rm(maps1d,indicesEffects(:,1),sujets);
             disp('ANOVA1rm')
-            testANOVA.name='ANOVA1rm';
+            anova.type='ANOVA1rm';
         end
     else
         ANOVA=[];
         disp('no ANOVA required')
-        testANOVA.name='no ANOVA';
+        anova.type='no ANOVA';
         indicesAnovaEffects{1}=1;
     end
 end
@@ -29,15 +29,15 @@ if nEffects==2
     if nRm==0
         ANOVA=spm1d.stats.nonparam.anova2(maps1d,indicesEffects(:,1),indicesEffects(:,2));
         disp('ANOVA2')
-        testANOVA.name='ANOVA2';
+        anova.type='ANOVA2';
     elseif nRm==1
         ANOVA=spm1d.stats.nonparam.anova2onerm(maps1d,indicesEffects(:,1),indicesEffects(:,2),sujets);
         disp('ANOVA2 1rm')
-        testANOVA.name='ANOVA2 1rm';
+        anova.type='ANOVA2 1rm';
     elseif nRm==2
         ANOVA=spm1d.stats.nonparam.anova2rm(maps1d,indicesEffects(:,1),indicesEffects(:,2),sujets);
         disp('ANOVA2rm')
-        testANOVA.name='ANOVA2rm';
+        anova.type='ANOVA2rm';
     end
     
     
@@ -53,32 +53,32 @@ elseif nEffects==3
     if nRm==0
         ANOVA=spm1d.stats.nonparam.anova3(maps1d,indicesEffects(:,1),indicesEffects(:,2),indicesEffects(:,3));
         disp('ANOVA3')
-        testANOVA.name='ANOVA3';
+        anova.type='ANOVA3';
     elseif nRm==1
         ANOVA=spm1d.stats.nonparam.anova3onerm(maps1d,indicesEffects(:,1),indicesEffects(:,2),indicesEffects(:,3),sujets);
         disp('ANOVA3 1rm')
-        testANOVA.name='ANOVA3 1rm';
+        anova.type='ANOVA3 1rm';
     elseif nRm==2
         ANOVA=spm1d.stats.nonparam.anova3tworm(maps1d,indicesEffects(:,1),indicesEffects(:,2),indicesEffects(:,3),sujets);
         disp('ANOVA3 2rm')
-        testANOVA.name='ANOVA3 2rm';
+        anova.type='ANOVA3 2rm';
     elseif nRm==3
         ANOVA=spm1d.stats.nonparam.anova3rm(maps1d,indicesEffects(:,1),indicesEffects(:,2),indicesEffects(:,3),sujets);
         disp('ANOVA3rm')
-        testANOVA.name='ANOVA3rm';
+        anova.type='ANOVA3rm';
     end
 end
 
 if nEffects==1
-    namesEffect=eNames{indicesAnovaEffects{1}};
[…]
