+++ src/fctPostHoc1d.m
@@ -22,7 +22,7 @@ if nEffects==1
             
             meansData{i}=maps1d(indicesEffects==combi{i}(1),:);
             legendPlot=[legendPlot,{char(modalitiesAll{1}(combi{i}(1)))}];
-            namesConditions{i}=[char(modalitiesAll{1}(combi{i}(1)))];
+            posthoc.data.names{i}=[char(modalitiesAll{1}(combi{i}(1)))];
             
         end
         
@@ -34,7 +34,7 @@ if nEffects==1
         savefig([savedir '/FIG/' verifSaveName(eNames{1})])
         close
         
-        mapsConditions=meansData;
+        posthoc.data.continuum=meansData;
         clear meansData
         
         loop=0;
@@ -59,7 +59,7 @@ if nEffects==1
             for i=1:2
                 % comparison + name
                 DATA{i}=maps1d(indicesEffects==combi{Comp{comp}(i)}(1),:);
-                namesDifferences{comp,i}=[char(modalitiesAll{1}(combi{Comp{comp}(i)}(1)))];
+                posthoc.differences.names{i,comp}=[char(modalitiesAll{1}(combi{Comp{comp}(i)}(1)))];
             end
             
             % t-test
@@ -67,9 +67,12 @@ if nEffects==1
                 differencesData{1}=DATA{1}-DATA{2};
                 relativeDifferencesData{1}=100*(DATA{1}-DATA{2})./DATA{2};
                 
+                posthoc.differences.continuum{1,comp}=differencesData{1};
+                posthoc.differences.continuumRelative{1,comp}=relativeDifferencesData{1};
+                
                 Ttest=spm1d.stats.nonparam.ttest_paired(DATA{1},DATA{2});
-                testTtests.name{comp}='paired';
-                [ES{comp},ESsd{comp}]=esCalculation(DATA);
+                posthoc.tTests.type{comp}='paired';
+                [posthoc.differences.ES{comp},posthoc.differences.ESsd{comp}]=esCalculation(DATA);
                 
             else
                 
@@ -77,82 +80,89 @@ if nEffects==1
                 differencesData{1}=mean(DATA{1})-mean(DATA{2});
                 relativeDifferencesData{1}=100*(mean(DATA{1})-mean(DATA{2}))./mean(DATA{2});
                 
+                posthoc.differences.continuum{1,comp}=differencesData{1};
+                posthoc.differences.continuumRelative{1,comp}=relativeDifferencesData{1};
+                
                 Ttest=spm1d.stats.nonparam.ttest2(DATA{1},DATA{2});
-                testTtests.name{comp}='independant';
-                [ES{comp},ESsd{comp}]=esCalculation(DATA);
+                posthoc.tTests.type{comp}='independant';
+                [posthoc.differences.ES{comp},posthoc.differences.ESsd{comp}]=esCalculation(DATA);
             end
             
-            mapsDifferences{1,comp}=differencesData{1};
-            mapsDifferences{2,comp}=relativeDifferencesData{1};
             
             % inference
-            [testTtests.nWarning{comp},iterations,alpha]=fctWarningIterations(Ttest,alphaOriginal,multiIterations,maximalIT,IT);
-            testTtests.alphaOriginal{comp}=alphaOriginal;
-            testTtests.alpha{comp}=alpha;
-            testTtests.nIterations{comp}=iterations;
[…]
