### 2021-04-14 47485f6
Reduce tiff fig
default option is now 720*480p at 96ppp.
