+++ src/plot/displayTtest.m
@@ -9,7 +9,7 @@
 function []=displayTtest(mapT,Tthreshold,anovaEffects,Fs,xlab,ylab,ylimits,dimensions,nx,ny,xlimits,imageFontSize,imageSize,colorMap)
 
 if isempty(imageSize)
-    figure('Units', 'Normalized', 'OuterPosition', [0, 0, 1, 1],'visible','off');
+    figure('Units', 'Pixels', 'OuterPosition', [0, 0, 720, 480],'visible','off');
 elseif max(size(imageSize))==1
     figure('Units', 'Centimeter', 'OuterPosition', [0, 0, imageSize, imageSize],'visible','off');
 else
