+++ src/plot/displayRelativeDiffMaps.m
@@ -5,7 +5,7 @@
 function []=displayRelativeDiffMaps(map,Fs,xlab,ylab,ylimits,nx,ny,xlimits,imageFontSize,imageSize,colorMap,relativeRatio)
 
 if isempty(imageSize)
-    figure('Units', 'Normalized', 'OuterPosition', [0, 0, 1, 1],'visible','off');
+    figure('Units', 'Pixels', 'OuterPosition', [0, 0, 720, 480],'visible','off');
 elseif max(size(imageSize))==1
     figure('Units', 'Centimeter', 'OuterPosition', [0, 0, imageSize, imageSize],'visible','off');
 else
