### 2021-04-14 11113b8
Update fctSPM.m
