+++ Article@JOSS/paper.md
@@ -28,7 +28,8 @@ As in "classical" statistics on scalar values (0D), there is a parametric and a 
 # Statement of need
 Most of physiological data measured during human movement are continuous and expressed in function of time. However, researchers predominantly analyze extracted scalar values from the continuous measurement, as the mean, the maximum, the amplitude, or the integrated value over the time. Analyzing continuous values (i.e., time series) can provide more information than extracted indicators, as the later discards one dimension of the data among the magnitude and localization in time. In addition, oscillatory signals such as muscle vibrations and electromyograms contain information in the temporal and frequency domains. However, scalar analysis reduces the information at only one dimension by discarding two dimensions among the magnitude and the localization in the time and/or frequency domain.
 
-To analyze all the dimensions of a signal without losing information, the analysis of curves or maps was proposed, coded, and put online by Pataky (https://github.com/0todd0000/spm1dmatlab). However, the use of the proposed functions does not allow the analysis of 2D data automatically. Moreover, a rather frequent error is to consider only the significance of the last statistical test performed and not the intersection between the post-hoc tests and the ANOVA. Indeed, a difference between two samples can be significant if, and only if the ANOVA is significant in the same areas.
+To analyze all the dimensions of a signal without losing information, the analysis of curves or maps was proposed, coded, and put online by Pataky. However, the use of the proposed functions does not allow the analysis of 2D data automatically. Moreover, a rather frequent error is to consider only the significance of the last statistical test performed and not the intersection between the post-hoc tests and the ANOVA. Indeed, a difference between two samples can be significant if, and only if the ANOVA is significant in the same areas.
+This package, integrated in the fctSPM repository at `./fctSPM/src/spm1d_Pataky` is [published elsewhere](https://github.com/0todd0000/spm1dmatlab), and thus is not part of this JOSS review.
 
 # fctSPM
 The function we propose meets two objectives. 1/ to allow statistical inferences on curves and maps with a standardized format and 2/ to simplify analyses by comparing means while considering intersections with tests performed upstream (ANOVA and post-hoc of main effects).
