+++ Article@JOSS/paper.bib
@@ -35,7 +35,7 @@
   title   = {{Is accelerometry an effective method to assess muscle vibrations in comparison to ultrafast ultrasonography}},
   journal = {IEEE Transactions on Biomedical Engineering},
   doi = {10.1109/TBME.2020.3035838},
-  year    = 2020,
-  volume={},
-  pages={},
+  year    = 2021,
+  volume=68,
+  pages={1409-1416},
 }
