### 2021-04-14 a21d126
Update chooseSavedir.m
when choose to not erase files, the default search path is the previously choosen directory
