+++ Examples/D1_independantTtest.m
@@ -1,12 +1,10 @@
-% This function takes ~10 seconds on i5 processor and files takes 20 Mo of
+% This function takes ~15 seconds on i5 processor and files takes 1 Mo of
 % storage
+% don't forget to add the source code path
 
 clear
 close all
 clc
-
-% add source code path
-addpath(genpath("../src"))
 
 % data
 load ExampleDatas
