+++ src/fctPostHoc1d.m
@@ -105,7 +105,7 @@ if nEffects==1
             clustersT=extractClusterData(Ttest_inf.clusters);
             for c=1:numel(clustersT)
                 posthoc.tTests.clusterLocation{comp}{c}=clustersT{c}.endpoints;
-                posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P;
+                posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*nComp;
             end
             
             plotmean(differencesData,IC,xlab,ylab,Fs,xlimits,nx,[],[],imageFontSize,imageSize,transparancy1D,[])
@@ -286,7 +286,7 @@ if nEffects==2
                 clustersT=extractClusterData(Ttest_inf.clusters);
                 for c=1:numel(clustersT)
                     posthoc.tTests.clusterLocation{comp}{c}=clustersT{c}.endpoints;
-                    posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P;
+                    posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*nComp;
                 end
                 % plot of spm analysis
                 displayTtest(posthoc.tTests.Tcontinuum{1,comp},posthoc.tTests.Tthreshold{comp},posthoc.tTests.Tsignificant{1,comp},Fs,xlab,ylab,ylimits,dimensions,nx,ny,xlimits,imageFontSize,imageSize,transparancy1D)
@@ -454,7 +454,7 @@ if nEffects==3
                 clustersT=extractClusterData(Ttest_inf.clusters);
                 for c=1:numel(clustersT)
                     posthoc.tTests.clusterLocation{comp}{c}=clustersT{c}.endpoints;
-                    posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P;
+                    posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*nComp;
                 end
                 
                 
@@ -640,7 +640,7 @@ if nEffects==3
                 clustersT=extractClusterData(Ttest_inf.clusters);
                 for c=1:numel(clustersT)
                     posthoc.tTests.clusterLocation{comp}{c}=clustersT{c}.endpoints;
-                    posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P;
+                    posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*nComp;
                 end
                 
                 
@@ -915,7 +915,7 @@ if nEffects>1
             clustersT=extractClusterData(Ttest_inf.clusters);
             for c=1:numel(clustersT)
                 posthoc.tTests.clusterLocation{comp}{c}=clustersT{c}.endpoints;
-                posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P;
+                posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*nComp;
             end
             
             if nEffects==2
