### 2021-04-15 e630117
alpha risk propagation
alpha risk defined for the anova is propagated to post hoc tests.

nT option was deleted.

alphaT option was set as the same as alpha by default.
