### 2021-04-15 d81053d
Update fctAnova.m
