+++ src/fctAnova.m
@@ -36,7 +36,7 @@ if ~ignoreAnova
             if min(dimensions)==1
                 for c=1:numel(clustersAnova)
                     anova.clusterLocation{c}=clustersAnova{c}.endpoints;
-                    anova.clusterP(c)=clustersAnova{c}.P*(0.05/alpha);;
+                    anova.clusterP(c)=clustersAnova{c}.P*(0.05/alpha);
                 end
             end
             
