+++ src/fctPostHoc1d.m
@@ -106,7 +106,7 @@ if nEffects==1
             clustersT=extractClusterData(Ttest_inf.clusters);
             for c=1:numel(clustersT)
                 posthoc.tTests.clusterLocation{comp}{c}=clustersT{c}.endpoints;
-                posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*nComp;
+                posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*(0.05/alpha);
             end
             
             plotmean(differencesData,IC,xlab,ylab,Fs,xlimits,nx,[],[],imageFontSize,imageSize,transparancy1D,[])
@@ -283,7 +283,7 @@ if nEffects==2
                 clustersT=extractClusterData(Ttest_inf.clusters);
                 for c=1:numel(clustersT)
                     posthoc.tTests.clusterLocation{comp}{c}=clustersT{c}.endpoints;
-                    posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*nComp;
+                    posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*(0.05/alpha);
                 end
                 % plot of spm analysis
                 displayTtest(posthoc.tTests.Tcontinuum{1,comp},posthoc.tTests.Tthreshold{comp},posthoc.tTests.Tsignificant{1,comp},Fs,xlab,ylab,ylimits,dimensions,nx,ny,xlimits,imageFontSize,imageSize,transparancy1D)
@@ -447,7 +447,7 @@ if nEffects==3
                 clustersT=extractClusterData(Ttest_inf.clusters);
                 for c=1:numel(clustersT)
                     posthoc.tTests.clusterLocation{comp}{c}=clustersT{c}.endpoints;
-                    posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*nComp;
+                    posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*(0.05/alpha);
                 end
                 
                 
@@ -629,7 +629,7 @@ if nEffects==3
                 clustersT=extractClusterData(Ttest_inf.clusters);
                 for c=1:numel(clustersT)
                     posthoc.tTests.clusterLocation{comp}{c}=clustersT{c}.endpoints;
-                    posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*nComp;
+                    posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*(0.05/alpha);
                 end
                 
                 
@@ -900,7 +900,7 @@ if nEffects>1
             clustersT=extractClusterData(Ttest_inf.clusters);
             for c=1:numel(clustersT)
                 posthoc.tTests.clusterLocation{comp}{c}=clustersT{c}.endpoints;
-                posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*nComp;
+                posthoc.tTests.clusterP{comp}(c)=clustersT{c}.P*(0.05/alpha);
             end
             
             if nEffects==2
