+++ Examples/D2_pairedTtest.m
@@ -12,8 +12,8 @@ load ExampleDatas
 DATA=ExampleDatas.BAB(:,1:2);
 
 % parameters
-EFFET_ind=[];
-EFFET_rm{1}={'C1','C2'};
+independantEffects=[];
+repeatedMeasuresEffects{1}={'C1','C2'};
 effectNames={'Shoes'};
 % There are 9 subjects
 % Data(:,1) correspond to Shoes=C1
@@ -29,7 +29,7 @@ ylimits=[15 200];
 
 % SPM
 tic
-fctSPM(DATA,EFFET_ind,EFFET_rm,...
+fctSPM(DATA,independantEffects,repeatedMeasuresEffects,...
     'savedir',savedir,...
     'effectsNames',effectNames,...
     'xlabel',xlab,'ylabel',ylab,...
