+++ Examples/D2_ANOVA1.m
@@ -13,8 +13,8 @@ load ExampleDatas
 DATA=ExampleDatas.ACCELECHO;
 
 % parameters
-EFFET_ind{1}={'L','L','L','L','L','L','L','M','M','M','M','S','S','S','S'}; % same number than participants
-EFFET_rm=[];
+independantEffects{1}={'L','L','L','L','L','L','L','M','M','M','M','S','S','S','S'}; % same number than participants
+repeatedMeasuresEffects=[];
 effectNames={'Group'};
 % There is 15 subjects
 % Subjects 1 to 7 are 'L', 8 to 11 are 'M', and 12 to 15 are 'S'
@@ -31,7 +31,7 @@ ylimits=[10 130];
 
 % SPM
 tic
-fctSPM(DATA,EFFET_ind,EFFET_rm,...
+fctSPM(DATA,independantEffects,repeatedMeasuresEffects,...
     'savedir',savedir,...
     'effectsNames',effectNames,...
     'xlabel',xlab,'ylabel',ylab,...
