+++ Examples/D2_independantTtest.m
@@ -12,8 +12,8 @@ load ExampleDatas
 DATA=ExampleDatas.ACCELECHO;
 
 % parameters
-EFFET_ind{1}={'L','L','L','L','L','L','L','L','S','S','S','S','S','S','S'}; % same number than participants
-EFFET_rm=[];
+independantEffects{1}={'L','L','L','L','L','L','L','L','S','S','S','S','S','S','S'}; % same number than participants
+repeatedMeasuresEffects=[];
 effectNames={'Group'};
 % There is 15 subjects
 % Subjects 1 to 8 are 'L', and 9 to 15 are 'S'
@@ -28,7 +28,7 @@ ylimits=[10 130];
 
 % SPM
 tic
-fctSPM(DATA,EFFET_ind,EFFET_rm,...
+fctSPM(DATA,independantEffects,repeatedMeasuresEffects,...
     'savedir',savedir,...
     'effectsNames',effectNames,...
     'xlabel',xlab,'ylabel',ylab,...
