### 2021-04-20 28c56ee
update examples
add fctSPMS and saveNplot to examples
