+++ Examples/D2_ANOVA2rm.m
@@ -23,20 +23,36 @@ effectNames={'Device','Activation'};
 % Data(:,4) correspond to Device=US and Activation=Relaxed
 
 savedir=[];
+savedir2=[];
 xlab='Time (s)';
 ylab='Frequency (Hz)';
 Fs=500;
 colorbarLabel='Amplitude (m\cdots^-^2)';
 limitMeanMaps=12;
 ylimits=[10 130];
+nTicksY=7;
+nTicksX=6;
 
 % SPM
 tic
-fctSPM(DATA,independantEffects,repeatedMeasuresEffects,...
+spmAnalysis=fctSPM(DATA,independantEffects,repeatedMeasuresEffects,...
     'savedir',savedir,...
     'effectsNames',effectNames,...
-    'xlabel',xlab,'nTicksX',6,'ylabel',ylab,'nTicksY',7,...
+    'xlabel',xlab,'ylabel',ylab,...
     'sampleFrequency',Fs,'colorbarLabel',colorbarLabel,...
     'limitMeanMaps',limitMeanMaps,...
-    'ylimits',ylimits);
+    'ylimits',ylimits,'nTicksX',nTicksX,'nTicksY',nTicksY);
+toc
+
+tic
+spmAnalysis2=fctSPMS(DATA,independantEffects,repeatedMeasuresEffects,'effectsNames',effectNames);
+toc
+
+tic
+saveNplot(spmAnalysis2,...
+    'savedir',savedir2,...
+    'xlabel',xlab,'ylabel',ylab,...
+    'sampleFrequency',Fs,'colorbarLabel',colorbarLabel,...
+    'limitMeanMaps',limitMeanMaps,...
+    'ylimits',ylimits,'nTicksX',nTicksX,'nTicksY',nTicksY);
 toc
