+++ Examples/D1_ANOVA2_1rm.m
@@ -22,6 +22,7 @@ repeatedMeasuresEffects{1}={'Right','Left'};
 colorLine{2}=[rgb('green'); rgb('red')];
 
 savedir=[];
+savedir2=[];
 xlab='Angle (°)';
 ylab='Ratio';
 xlimits=[30 90];
@@ -31,11 +32,20 @@ nTicksX=7;
 
 % SPM
 tic
-fctSPM(DATA,independantEffects,repeatedMeasuresEffects,...
+spmAnalysis=fctSPM(DATA,independantEffects,repeatedMeasuresEffects,...
     'savedir',savedir,...
     'effectsNames',effectNames,...
-    'xlabel',xlab,'ylabel',ylab,...
-    'xlimits',xlimits,'ylimits',ylimits,'nTicksY',nTicksY,'nTicksX',nTicksX,...
-    'transparancy1D',0.05,'colorline',colorLine,'imageSize',15);
+    'xlabel',xlab,'ylabel',ylab,'nTicksX',nTicksX,...
+    'xlimits',xlimits,'ylimits',ylimits,'nTicksY',nTicksY,'colorline',colorLine);
 toc
 
+tic
+spmAnalysis2=fctSPMS(DATA,independantEffects,repeatedMeasuresEffects,'effectsNames',effectNames);
+toc
+
+tic
+saveNplot(spmAnalysis2,...
+    'savedir',savedir2,...
+    'xlabel',xlab,'ylabel',ylab,'nTicksX',nTicksX,...
+    'xlimits',xlimits,'ylimits',ylimits,'nTicksY',nTicksY);
+toc
