+++ README.md
@@ -5,7 +5,7 @@ The function automatically adapts to 1D and 2D data.
 
 The general usage is:
 ```matlab
-fctSPM(data, independantEffects, repeatedMeasuresEffects, varargin)
+spmAnalysis=fctSPM(data, independantEffects, repeatedMeasuresEffects, varargin)
 ```
 
 
@@ -17,8 +17,9 @@ fctSPM(data, independantEffects, repeatedMeasuresEffects, varargin)
 - [Outputs](#Outputs)
 - [Examples](#Examples)
 - [Using fctSPM](#Using-fctSPM)
-- [Obligatory INPUTS](#Obligatory-INPUTS)
-- [Optional INPUTS](#Optional-INPUTS)
+- [Obligatory inputs](#Obligatory-inputs)
+- [Optional inputs](#Optional-inputs)
+- [Optional functions](#Optional-functions)
 
 ## Warnings ##
 - Unbalanced two- and three-way repeated-measures ANOVA results have not been verified.
@@ -53,16 +54,11 @@ Compatible from Matlab R2017b
 
 
 ## Outputs ##
-This fonction creates two folders:
-* `ANOVA` (no ANOVA folder is created if the analysis is a t-test).  
-* `Post hoc` for the post-hoc analysis.  
 
-Each folder is composed of two different types of files : 
-* `.mat` files in which all data and statistical analysis can be found
-* `.TIF` and `.fig` files representing the statistical analysis on different graphics.  
+### spmAnalysis ###
+`spmAnalaysis.mat` is a structure composed of the results of teh statistical analysis (ANOVA + Post Hoc). 
 
-### ANOVA folder ###
-`anova.mat` file is composed of different fields (same for 1D and 2D) :  
+`spmAnalysis.anova` is composed of different fields :  
 * `type` is the type of ANOVA performed
 * `effectNames` is a structure (one cell for each effect) that represent the names of the effects tested (mains and interactions)
 * `alphaOriginal` is the alpha risk choosen for the anova (default is 0.05 (5%)).  
@@ -77,14 +73,7 @@ Each folder is composed of two different types of files :
 
 `clusterLocation` and `clusterP` are created only in one dimension
 
-Each effect is also display on a specific figure in `.TIF` format named after `effectNames`, a floder named FIG is also created and contains the figures in `.fig` format.   
-For 1D:  The curve represents the `Fcontinuum`, the horizontal line is the `Fthreshold`, the highlighted parts of the curve in blue represents the significant cluster(s), and the vertical lines are the start and end indexes.
-For 2D: The map represents the `Fcontinuum`, with a colorbar maximum at `Fthreshold`, the white clusters represent the significant clusters.
-
-
-### Post hoc folder ###
-This folder contains additional folders (0 (for anova1), 3 (for anova2) or 7 (for anova3)) that contain figures and metrics representing the different post-hoc tests.  
-`posthoc.mat` file is composed of different fields :
+`spmAnalysis.posthoc` is a strucure of cells (one for each effect of the ANOVA) composed of different fields :
 * `data.names` is a structure that contains the name of the conditions used in the analysis (\cap is the union of different conditions for interactions).
 * `data.continuum` is a structure that contains the data used in the analysis.
 
@@ -96,9 +85,11 @@ For the following outputs, the number of cells of the structure correspond to th
 * `differences.ESsd` is the standard deviation of the effect size.
 * `tTests.type` is the type of t-test performed (independant or paired)
 * `tTests.names` is the name of the conditions (the first minus the second) used in the differences and t-tests.
[…]
