### 2021-04-20 1b64362
S functions
add the *S functions to perform analysis without ploting or saving
