+++ src/fctPostHoc2dS.m
@@ -1,0 +1,664 @@
+function posthoc=fctPostHoc2dS(nEffects,indicesEffects,maps1d,dimensions,modalitiesAll,typeEffectsAll,eNames,multiIterations,IT,anovaEffects,maximalIT,doAllInteractions,alphaT,alphaAOV)
+
+%% define alpha risk
+if isempty(alphaT)
+    alphaT=alphaAOV;
+    warnT=0;
+else
+    if alphaT~=alphaAOV
+        warning('Post-hoc analysis is not valid. Please keep the same alpha risk for ANOVA and post hoc tests');
+        warnT=1;
+    else
+        warnT=0;
+    end
+end
+
+%% T-TEST 1 EFFECT = MAIN EFFECT
+if nEffects==1
+    
+    posthoc{1}.name=eNames{1};
+    
+    loop=0;
+    for i=1:max(indicesEffects)
+        loop=loop+1;
+        combi{loop}=i;
+    end
+    nCombi=size(combi,2);
+    
+    for i=1:nCombi
+        
+        % means
+        meansData=reshape(mean(maps1d(indicesEffects==combi{i}(1),:)),dimensions(1),dimensions(2));
+        stdData=reshape(std(maps1d(indicesEffects==combi{i}(1),:)),dimensions(1),dimensions(2));
+        posthoc{1}.data.names{i}=[char(modalitiesAll{1}(combi{i}(1)))];
+        posthoc{1}.data.meanContinuum{i}=meansData;
+        posthoc{1}.data.sdContinuum{i}=stdData;
+        
+    end
+    
+    loop=0;
+    for i=1:size(combi,2)
+        for j=1:size(combi,2)
+            if i<j
+                loop=loop+1;
+                Comp{loop}=[i j];
+            end
+        end
+    end
+    
+    nComp=size(Comp,2);
+    alphaBonferronni=alphaT/nComp;
+    
+    for comp=1:nComp
+        
+        for i=1:2
+            % comparison + name
+            DATA{i}=maps1d(indicesEffects==combi{Comp{comp}(i)}(1),:);
+        end
+        posthoc{1}.differences.names{1,comp}=char([modalitiesAll{1}{combi{Comp{comp}(1)}(1)} ' - ' modalitiesAll{1}{combi{Comp{comp}(2)}(1)}]);
+        
+        if typeEffectsAll>0
+            
+            % differences
+            differencesData=reshape(mean(DATA{1}-DATA{2}),dimensions(1),dimensions(2));
+            relativeDifferencesData=reshape(100*mean((DATA{1}-DATA{2})./DATA{2}),dimensions(1),dimensions(2));
+            
+            posthoc{1}.differences.continuum{1,comp}=differencesData;
+            posthoc{1}.differences.continuumRelative{1,comp}=relativeDifferencesData;
+            
+            Ttest=spm1d.stats.nonparam.ttest_paired(DATA{1},DATA{2});
+            posthoc{1}.tTests.type{comp}='paired';
+            [posthoc{1}.differences.ES{comp},posthoc{1}.differences.ESsd{comp}]=esCalculation(DATA);
+            
+        else
+            
+            % differences
+            differencesData=reshape(mean(DATA{1})-mean(DATA{2}),dimensions(1),dimensions(2));
+            relativeDifferencesData=reshape(100*(mean(DATA{1})-mean(DATA{2}))./mean(DATA{2}),dimensions(1),dimensions(2));
+            
+            posthoc{1}.differences.continuum{1,comp}=differencesData;
+            posthoc{1}.differences.continuumRelative{1,comp}=relativeDifferencesData;
+            
[…]
