+++ src/fctSPMS.m
@@ -1,0 +1,160 @@
+% Trama Robin (LIBM) 04/2021 --> JOSS
+% trama.robin@gmail.com
+
+% available at :
+% - https://github.com/tramarobin/fctSPM
+% - https://www.mathworks.com/matlabcentral/fileexchange/77945-fctspm
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+%% ¨Please read README.md on https://github.com/tramarobin/fctSPM for all informations
+
+% Using spm1d package (v.0.4.3), compute anova and post-hoc tests from anova1 to anova3rm, with a non-parametric approach (permutation tests)
+% The type of anova (if required) and post-hoc are choosen regarding the independant or repeated measure effect given in parameters.
+% The function automatically adapts to 1D and 2D data
+% Examples are in ...\fctSPM\Examples
+% 1D examples are torque ratios
+% 2D examples are maps obtained with continuous wavelet transforms
+
+% Please visit http://spm1d.org/index.html for information
+% spm1d package for matlab is published elswhere : https://github.com/0todd0000/spm1dmatlab
+
+% please cite for spm1d : Pataky TC (2010). Generalized n-dimensional biomechanical field analysis using statistical parametric mapping. Journal of Biomechanics 43, 1976-1982.
+% please cite for permutation tests : Nichols TE, Holmes AP (2002). Nonparametric permutation tests for functional neuroimaging: a primer with examples. Human Brain Mapping 15(1), 1–25.
+
+% WARNINGS:
+
+% Unbalanced two- and three-way repeated-measures ANOVA results have not been verified.
+% Example: more subjects in Group 1 than in Group 2.
+% Please interpret results for these designs with caution, and recognize that they might not be valid.
+
+% Post-hoc tests with Bonferonni correction are only approximate
+
+% Effect sizes and confidence intervals are not validated
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+% OUTPUTS
+%`spmAnalaysis.mat` is a structure composed of the results of teh statistical analysis (ANOVA + Post Hoc). 
+
+%`spmAnalysis.anova` is composed of different fields :  
+%* `type` is the type of ANOVA performed
+%* `effectNames` is a structure (one cell for each effect) that represent the names of the effects tested (mains and interactions)
+%* `alphaOriginal` is the alpha risk choosen for the anova (default is 0.05 (5%)).  
+%* `pCritical` is the alpha risk used for the anova. Warning message is displayed if this value is modified.
+%* `nIterations` is the number of iterations performed for the anova.
+%* `maxIterations` is the number of maximal iterations possible for the anova.
+%* `Fcontinuum` is a structure that represent the F-value for each node
+%* `Fthreshold` is a structure that represent the statistical threshold for the F-values (statistical inference)
+%* `Fsignificant` is a structure that contains the logical for the significance (Fcontinuum > Fthreshold) of each effect of the ANOVA (1 if significant, 0 if not).  
+%* `clusterLocation` is a structure (one for each significant cluster) that contains the location (start and end as indexes) of each significant cluster.
+%* `clusterP` is a structure (one for each significant cluster) that contains the p-value of each significant cluster. 
+
+%`clusterLocation` and `clusterP` are created only in one dimension
+
+%`spmAnalysis.posthoc` is a strucure of cells (one for each effect of the ANOVA) composed of different fields :
[…]

[…]
