### 2021-04-20 a1a3e1b
alpha warnings + outputs
warning are displayed if alphaT is modified
spmAnalysis is created (variable) and saved instead of several files
