+++ src/fctSPM.m
@@ -1,4 +1,4 @@
-% Trama Robin (LIBM) 30/03/2021 --> release 1.4.3
+% Trama Robin (LIBM) 04/2021 --> JOSS
 % trama.robin@gmail.com
 
 % available at :
@@ -12,6 +12,7 @@
 % Using spm1d package (v.0.4.3), compute anova and post-hoc tests from anova1 to anova3rm, with a non-parametric approach (permutation tests)
 % The type of anova (if required) and post-hoc are choosen regarding the independant or repeated measure effect given in parameters.
 % The function automatically adapts to 1D and 2D data
+% Analysis and figures of the analysis are saved
 % Examples are in ...\fctSPM\Examples
 % 1D examples are torque ratios
 % 2D examples are maps obtained with continuous wavelet transforms
@@ -35,91 +36,91 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 % OUTPUTS
-% This fonction creates two folders:
-% * `ANOVA` (no ANOVA folder is created if the analysis is a t-test).
-% * `Post hoc` for the post-hoc analysis.
+%`spmAnalaysis.mat` is a structure composed of the results of teh statistical analysis (ANOVA + Post Hoc). 
 
-% Each folder is composed of two different types of files :
-% * `.mat` files in which all data and statistical analysis can be found
-% * `.TIF` and `.fig` files representing the statistical analysis on different graphics.
+%`spmAnalysis.anova` is composed of different fields :  
+%* `type` is the type of ANOVA performed
+%* `effectNames` is a structure (one cell for each effect) that represent the names of the effects tested (mains and interactions)
+%* `alphaOriginal` is the alpha risk choosen for the anova (default is 0.05 (5%)).  
+%* `pCritical` is the alpha risk used for the anova. Warning message is displayed if this value is modified.
+%* `nIterations` is the number of iterations performed for the anova.
+%* `maxIterations` is the number of maximal iterations possible for the anova.
+%* `Fcontinuum` is a structure that represent the F-value for each node
+%* `Fthreshold` is a structure that represent the statistical threshold for the F-values (statistical inference)
+%* `Fsignificant` is a structure that contains the logical for the significance (Fcontinuum > Fthreshold) of each effect of the ANOVA (1 if significant, 0 if not).  
+%* `clusterLocation` is a structure (one for each significant cluster) that contains the location (start and end as indexes) of each significant cluster.
+%* `clusterP` is a structure (one for each significant cluster) that contains the p-value of each significant cluster. 
 
-% ### ANOVA folder ###
-% `anova.mat` file is composed of different fields (same for 1D and 2D) :
-% * `type` is the type of ANOVA performed
-% * `effectNames` is a structure (one cell for each effect) that represent the names of the effects tested (mains and interactions)
-% * `alphaOriginal` is the alpha risk choosen for the anova (default is 0.05 (5%)).
-% * `pCritical` is the alpha risk used for the anova.  Warning message is displayed if this value is modified.
-% * `nIterations` is the number of iterations performed for the anova.
-% * `maxIterations` is the number of maximal iterations possible for the anova.
-% * `Fcontinuum` is a structure that represent the F-value for each node
-% * `Fthreshold` is a structure that represent the statistical threshold for the F-values (statistical inference)
-% * `Fsignificant` is a structure that contains the logical for the significance (Fcontinuum > Fthreshold) of each effect of the ANOVA (1 if significant, 0 if not).
-% * `clusterLocation` is a structure (one for each significant cluster) that contains the location (start and end as indexes) of each significant cluster.
[…]
