+++ src/fctPostHoc1d.m
@@ -1,4 +1,4 @@
-function []=fctPostHoc1d(nEffects,indicesEffects,maps1d,dimensions,modalitiesAll,typeEffectsAll,eNames,savedir,multiIterations,IT,xlab,ylab,Fs,imageResolution,IC,ylimits,nx,ny,xlimits,anovaEffects,maximalIT,colorLine,doAllInteractions,imageFontSize,imageSize,alphaT,alphaAOV,transparancy1D,ratioSPM,yLimitES,spmPos,aovColor)
+function posthoc=fctPostHoc1d(nEffects,indicesEffects,maps1d,dimensions,modalitiesAll,typeEffectsAll,eNames,savedir,multiIterations,IT,xlab,ylab,Fs,imageResolution,IC,ylimits,nx,ny,xlimits,anovaEffects,maximalIT,colorLine,doAllInteractions,imageFontSize,imageSize,alphaT,alphaAOV,transparancy1D,ratioSPM,yLimitES,spmPos,aovColor)
 close all
 set(0, 'DefaultFigureVisible', 'off');
 savedir=[savedir '/Post hoc/'];
@@ -6,6 +6,14 @@ savedir=[savedir '/Post hoc/'];
 %% define alpha risk
 if isempty(alphaT)
     alphaT=alphaAOV;
+    warnT=0;
+else
+    if alphaT~=alphaAOV
+        warning('Post-hoc analysis is not valid. Please keep the same alpha risk for ANOVA and post hoc tests');
+        warnT=1;
+    else
+        warnT=0;
+    end
 end
 
 %% T-TEST 1 EFFECT = MAIN EFFECT
@@ -14,6 +22,7 @@ if nEffects==1
     if max(anovaEffects{1})==1 | doAllInteractions==1
         
         createSavedir(savedir)
+        posthoc{1}.name=eNames{1};
         
         loop=0;
         for i=1:max(indicesEffects)
@@ -27,7 +36,7 @@ if nEffects==1
             
             meansData{i}=maps1d(indicesEffects==combi{i}(1),:);
             legendPlot=[legendPlot,{char(modalitiesAll{1}(combi{i}(1)))}];
-            posthoc.data.names{i}=[char(modalitiesAll{1}(combi{i}(1)))];
+            posthoc{1}.data.names{i}=[char(modalitiesAll{1}(combi{i}(1)))];
             
         end
         
@@ -39,7 +48,7 @@ if nEffects==1
         savefig([savedir '/FIG/' verifSaveName(eNames{1})])
         close
         
-        posthoc.data.continuum=meansData;
+        posthoc{1}.data.continuum=meansData;
         clear meansData
         
         loop=0;
@@ -53,14 +62,14 @@ if nEffects==1
         end
         
         nComp=size(Comp,2);
-        alphaOriginal=alphaT/nComp;
+        alphaBonferronni=alphaT/nComp;
         
         for comp=1:nComp
             
             for i=1:2
                 % comparison + name
                 DATA{i}=maps1d(indicesEffects==combi{Comp{comp}(i)}(1),:);
-                posthoc.differences.names{i,comp}=[char(modalitiesAll{1}(combi{Comp{comp}(i)}(1)))];
+                posthoc{1}.differences.names{i,comp}=[char(modalitiesAll{1}(combi{Comp{comp}(i)}(1)))];
             end
             
             % t-test
@@ -68,12 +77,12 @@ if nEffects==1
                 differencesData{1}=DATA{1}-DATA{2};
                 relativeDifferencesData{1}=100*(DATA{1}-DATA{2})./DATA{2};
                 
-                posthoc.differences.continuum{1,comp}=differencesData{1};
-                posthoc.differences.continuumRelative{1,comp}=relativeDifferencesData{1};
[…]

[…]
