### 2021-04-20 301398c
update variable names
alpha and pBonferroni instead of alphaOriginal and alphaBonferroni
