+++ src/fctPostHoc2dS.m
@@ -47,7 +47,7 @@ if nEffects==1
     end
     
     nComp=size(Comp,2);
-    alphaBonferronni=alphaT/nComp;
+    pBonferroni=alphaT/nComp;
     
     for comp=1:nComp
         
@@ -89,12 +89,12 @@ if nEffects==1
         
         % inference
         posthoc{1}.tTests.names=posthoc{1}.differences.names;
-        [posthoc{1}.tTests.nWarning{comp},iterations,alpha]=fctWarningIterations(Ttest,alphaBonferronni,multiIterations,maximalIT,IT);
-        posthoc{1}.tTests.alphaOriginal=alphaT;
+        [posthoc{1}.tTests.nWarning{comp},iterations,alpha]=fctWarningIterations(Ttest,pBonferroni,multiIterations,maximalIT,IT);
+        posthoc{1}.tTests.alpha=alphaT;
         if warnT==1
-            posthoc{1}.tTests.warning="alphaOriginal is not valid";
+            posthoc{1}.tTests.warning="alpha is not valid";
         end
-        posthoc{1}.tTests.alphaBonferronni=alphaBonferronni;
+        posthoc{1}.tTests.pBonferroni=pBonferroni;
         posthoc{1}.tTests.pCritical{comp}=alpha;
         posthoc{1}.tTests.nIterations{comp}=iterations;
         Ttest_inf=Ttest.inference(alpha,'iterations',iterations,'force_iterations',logical(1),'two_tailed',logical(1));
@@ -159,10 +159,10 @@ if nEffects==2
         end
         
         nComp=size(Comp,2);
-        alphaBonferronni=alphaT/nComp;
+        pBonferroni=alphaT/nComp;
         
         for comp=1:nComp
-            alphaBonferronni=alphaT/nComp;
+            pBonferroni=alphaT/nComp;
             for i=1:2
                 % comparison + name
                 DATA{i}=maps1d(indicesEffects(:,mainEffect(1))==combi{Comp{comp}(i)}(1),:);
@@ -201,12 +201,12 @@ if nEffects==2
             posthoc{mainEffect(1)}.differences.ESsd{comp}=reshape(posthoc{mainEffect(1)}.differences.ESsd{comp},dimensions(1),dimensions(2));
             
             % inference
-            [posthoc{mainEffect(1)}.tTests.nWarning{comp},iterations,alpha]=fctWarningIterations(Ttest,alphaBonferronni,multiIterations,maximalIT,IT);
-            posthoc{mainEffect(1)}.tTests.alphaOriginal=alphaT;
+            [posthoc{mainEffect(1)}.tTests.nWarning{comp},iterations,alpha]=fctWarningIterations(Ttest,pBonferroni,multiIterations,maximalIT,IT);
+            posthoc{mainEffect(1)}.tTests.alpha=alphaT;
             if warnT==1
-                posthoc{mainEffect(1)}.tTests.warning="alphaOriginal is not valid";
+                posthoc{mainEffect(1)}.tTests.warning="alpha is not valid";
             end
-            posthoc{mainEffect(1)}.tTests.alphaBonferronni=alphaBonferronni;
+            posthoc{mainEffect(1)}.tTests.pBonferroni=pBonferroni;
             posthoc{mainEffect(1)}.tTests.pCritical{comp}=alpha;
             posthoc{mainEffect(1)}.tTests.nIterations{comp}=iterations;
             Ttest_inf=Ttest.inference(alpha,'iterations',iterations,'force_iterations',logical(1),'two_tailed',logical(1));
@@ -273,10 +273,10 @@ if nEffects==3
         end
         
         nComp=size(Comp,2);
-        alphaBonferronni=alphaT/nComp;
+        pBonferroni=alphaT/nComp;
         
[…]
