+++ src/fctPostHoc1d.m
@@ -62,7 +62,7 @@ if nEffects==1
         end
         
         nComp=size(Comp,2);
-        alphaBonferronni=alphaT/nComp;
+        pBonferroni=alphaT/nComp;
         
         for comp=1:nComp
             
@@ -101,12 +101,12 @@ if nEffects==1
             
             % inference
             posthoc{1}.tTests.names=posthoc{1}.differences.names;
-            [posthoc{1}.tTests.nWarning{comp},iterations,alpha]=fctWarningIterations(Ttest,alphaBonferronni,multiIterations,maximalIT,IT);
-            posthoc{1}.tTests.alphaOriginal=alphaT;
+            [posthoc{1}.tTests.nWarning{comp},iterations,alpha]=fctWarningIterations(Ttest,pBonferroni,multiIterations,maximalIT,IT);
+            posthoc{1}.tTests.alpha=alphaT;
             if warnT==1
-                posthoc{1}.tTests.warning="alphaOriginal is not valid";
+                posthoc{1}.tTests.warning="alpha is not valid";
             end
-            posthoc{1}.tTests.alphaBonferronni=alphaBonferronni;
+            posthoc{1}.tTests.pBonferroni=pBonferroni;
             posthoc{1}.tTests.pCritical{comp}=alpha;
             posthoc{1}.tTests.nIterations{comp}=iterations;
             Ttest_inf=Ttest.inference(alpha,'iterations',iterations,'force_iterations',logical(1),'two_tailed',logical(1));
@@ -232,7 +232,7 @@ if nEffects==2
             end
             
             nComp=size(Comp,2);
-            alphaBonferronni=alphaT/nComp;
+            pBonferroni=alphaT/nComp;
             
             for comp=1:nComp
                 
@@ -282,12 +282,12 @@ if nEffects==2
                 
                 % inference
                 posthoc{mainEffect(1)}.tTests.names=posthoc{mainEffect(1)}.differences.names;
-                [posthoc{mainEffect(1)}.tTests.nWarning{comp},iterations,alpha]=fctWarningIterations(Ttest,alphaBonferronni,multiIterations,maximalIT,IT);
-                posthoc{mainEffect(1)}.tTests.alphaOriginal=alphaT;
+                [posthoc{mainEffect(1)}.tTests.nWarning{comp},iterations,alpha]=fctWarningIterations(Ttest,pBonferroni,multiIterations,maximalIT,IT);
+                posthoc{mainEffect(1)}.tTests.alpha=alphaT;
                 if warnT==1
-                    posthoc{mainEffect(1)}.tTests.warning="alphaOriginal is not valid";
+                    posthoc{mainEffect(1)}.tTests.warning="alpha is not valid";
                 end
-                posthoc{mainEffect(1)}.tTests.alphaBonferronni=alphaBonferronni;
+                posthoc{mainEffect(1)}.tTests.pBonferroni=pBonferroni;
                 posthoc{mainEffect(1)}.tTests.pCritical{comp}=alpha;
                 posthoc{mainEffect(1)}.tTests.nIterations{comp}=iterations;
                 Ttest_inf=Ttest.inference(alpha,'iterations',iterations,'force_iterations',logical(1),'two_tailed',logical(1));
@@ -398,7 +398,7 @@ if nEffects==3
             end
             
             nComp=size(Comp,2);
-            alphaBonferronni=alphaT/nComp;
+            pBonferroni=alphaT/nComp;
             
             for comp=1:nComp
                 
@@ -450,12 +450,12 @@ if nEffects==3
                 
                 % inference
[…]
