### 2021-04-20 e016601
Update fctSPM.m
