### 2021-04-20 b4da3bc
Update fctSPMS.m
