### 2021-04-20 #3159 @0todd0000
@cMadan I have completed the main portion of my review using repository issues, and @tramarobin has satisfactorily responded to all of my concerns. Before I recommend "accept" just a few minor additions are needed, as outlined below. These additions are JOSS requirements, and not directly related to the repository, so I have raised this matter here rather than in the repository.

* **Documentation**:  the "Statement of need" and "Community guidelines" sections appear to be absent from the documentation.
* **Software paper**:  there is a "References" section, but no references appear to be present.

@tramarobin: please add these.
