+++ README.md
@@ -11,7 +11,7 @@ spmAnalysis=fctSPM(data, independantEffects, repeatedMeasuresEffects, varargin)
 
 ## Table of contents ##
 - [Warnings](#Warnings)
-- [Why fctSPM](#Why-fctSPM)
+- [Statement of need](#Statement-of-need)
 - [Citing fctSPM](#Citing-fctSPM)
 - [MATLAB Release Compatibility](#MATLAB-Release-Compatibility)
 - [Outputs](#Outputs)
@@ -20,6 +20,8 @@ spmAnalysis=fctSPM(data, independantEffects, repeatedMeasuresEffects, varargin)
 - [Obligatory inputs](#Obligatory-inputs)
 - [Optional inputs](#Optional-inputs)
 - [Optional functions](#Optional-functions)
+- [Community guidelines](#Community-guidelines)
+
 
 ## Warnings ##
 - Unbalanced two- and three-way repeated-measures ANOVA results have not been verified.
@@ -30,17 +32,26 @@ Please interpret results for these designs with caution, and recognize that they
 - Effect sizes and confidence intervals are not validated and are just given as supplementary information.
 - For more information, please visit : https://spm1d.org/Documentation.html
 
+
 ### Caution ###
 - Avoid the same typo (e.g., POST, POST2) for the effect names, it affects the recognition for the multiples comparisons 
 - Avoid underscore (_) or minus (-) sign. Spaces are OK
 - Use '/' when selecting a saving directory (`savedir` option)
 - Once you are ok with the created figure, increase the number of iterations (`multiIT` or `IT` options) to achieve numerical stability. 1000 iterations are a good start but if your hardware allows it, the more is the better
 
-## Why fctSPM ##
+
+## Statement of need ##
+Most of physiological data measured during human movement are continuous and expressed in function of time. 
+However, researchers predominantly analyze extracted scalar values from the continuous measurement, as the mean, the maximum, the amplitude, or the integrated value over the time. 
+Analyzing continuous values (i.e., time series) can provide more information than extracted indicators, as the later discards one dimension of the data among the magnitude and localization in time. 
+In addition, oscillatory signals such as muscle vibrations and electromyograms contain information in the temporal and frequency domains. 
+Once again, scalar analysis reduces the information at only one dimension by discarding two dimensions among the magnitude and the localization in the time and/or frequency domain.
+
 fctSPM allows MATLAB users to create figures of 1D and 2D SPM analysis for ANOVA and post-hoc designs.    
 The statistical analysis is also saved in .mat files.   
 This function synthetises the main and interaction effects to display only the significant post-hoc regarding the results of the ANOVA.   
 For post-hoc for interaction effects, the main effect is also displayed if located elsewhere than the interaction effect.
+
 
 ## Citing fctSPM ##
 - This function : under review in JOSS   
@@ -128,7 +139,6 @@ Subfolders : Contains the pairewise comparison results
 * SPM: Tcontinuum and statistical inferences plots. Bold blue lines are located at the significant differences (corrected with the ANOVA).
 * FIG folder contains the above mentionned folder with the figures in `.fig` format.
 
-
 ##### In two dimensions #####
 Means maps for each condition are represented in one figure each. 
 The global effect of the post hoc precedure is display on a figure with the name of the effect. Mean maps are represented on the diagonal, pairewise differences on the top-right panel, and pairewise spm inferences on the bottom-left panel. 
@@ -187,9 +197,7 @@ Refer to the [MATLAB documentation regarding search paths](https://fr.mathworks.
[…]
