### 2021-04-21 #3159 @tramarobin
@0todd0000 thank you for having take the time to test and suggest modification for this function.

> Documentation: the "Statement of need" and "Community guidelines" sections appear to be absent from the documentation.
The documentation was modified accordingly.

> Software paper: there is a "References" section, but no references appear to be present.
On the generated pdf above the References section appears correctly.
