### 2021-04-21 1b54e5a
update for saveNplot
Display the means + SPM main effects for the main effects.
work in progress for the ineractions
