+++ src/fctAnova.m
@@ -31,7 +31,7 @@ if ~ignoreAnova
             anova.maxIterations=ANOVA_inf.nPermUnique;
             anova.Fcontinuum=reshape(ANOVA_inf.z,dimensions(1),dimensions(2));
             anova.Fthreshold=ANOVA_inf.zstar;
-            anova.Fsignificant=reshape(ANOVA_inf.z>=anova.Fthreshold,dimensions(1),dimensions(2));
+            anova.Fsignificant{1}=reshape(ANOVA_inf.z>=anova.Fthreshold,dimensions(1),dimensions(2));
             clustersAnova=extractClusterData(ANOVA_inf.clusters);
             if min(dimensions)==1
                 for c=1:numel(clustersAnova)
@@ -41,7 +41,7 @@ if ~ignoreAnova
             end
             
             % Plot of Anova Results
-            displayAnova(anova.Fcontinuum,anova.Fthreshold,anova.Fsignificant,Fs,xlab,ylab,ylimits,dimensions,nx,ny,xlimits,colorMap,imageSize,imageFontSize)
+            displayAnova(anova.Fcontinuum,anova.Fthreshold,anova.Fsignificant{1},Fs,xlab,ylab,ylimits,dimensions,nx,ny,xlimits,colorMap,imageSize,imageFontSize)
             if displayContour & size(anova.Fcontinuum,2)>1
                 dispContour(anova.Fcontinuum,anova.Fthreshold,contourColor,dashedColor,transparency,lineWidth,linestyle)
             end
@@ -50,7 +50,7 @@ if ~ignoreAnova
             savefig([savedir '/ANOVA/FIG/' verifSaveName(anova.effectNames)])
             close
             
-            anovaEffects{1}(1,:)=anova.Fsignificant(:); % values saved for the interpretation of post-hoc tests
+            anovaEffects{1}(1,:)=anova.Fsignificant{1}(:); % values saved for the interpretation of post-hoc tests
             
         else % ANOVA2 & % ANOVA3
             
