+++ src/saveNplot.m
@@ -94,8 +94,12 @@ addParameter(p,'relativeRatio',[],@isnumeric) % the relative differences maps wi
 
 % 1d plot parameters
 addParameter(p,'CI',[],@isnumeric); % confidence interval is used instead of standard deviation (0.7-->0.999), 0 to display SEM
+addParameter(p,'colorLine',[]); % colorline for plots (default  is "lines") // rgb triplet, if in cell, apply each color to each effect (independant effect first)
 addParameter(p,'transparancy1D',0.10); % transparancy of SD for 1D plot
 addParameter(p,'yLimitES',[]); % y-axis limits for ES representation
+addParameter(p,'ratioSPM',[1 3]); % ratio of SPM subplot relative to total figure (default if 1/3 of the figure)
+addParameter(p,'spmPos',[]); % postion of spm plot, default is bottom, any value will set the position to up
+addParameter(p,'aovColor','k'); % color of anova on SPM plot (color or rgb)
 
 parse(p,varargin{:});
 
@@ -125,7 +129,10 @@ relativeRatio=p.Results.relativeRatio;
 linestyle=p.Results.linestyle;
 transparancy1D=p.Results.transparancy1D;
 yLimitES=p.Results.yLimitES;
-
+ratioSPM=p.Results.ratioSPM;
+spmPos=p.Results.spmPos;
+aovColor=p.Results.aovColor;
+colorLine=p.Results.colorLine;
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 close all
@@ -155,7 +162,7 @@ if spmAnalysis.anova.type~="no ANOVA"
     if ischar(anova.effectNames)
         
         % Plot of Anova Results
-        displayAnova(anova.Fcontinuum,anova.Fthreshold,anova.Fsignificant,Fs,xlab,ylab,ylimits,dimensions,nTicksX,nTicksY,xlimits,colorMap,imageSize,imageFontSize)
+        displayAnova(anova.Fcontinuum,anova.Fthreshold,anova.Fsignificant{1},Fs,xlab,ylab,ylimits,dimensions,nTicksX,nTicksY,xlimits,colorMap,imageSize,imageFontSize)
         if displayContour & size(anova.Fcontinuum,2)>1
             dispContour(anova.Fcontinuum,anova.Fthreshold,contourColor,dashedColor,transparancy,lineWidth,linestyle)
         end
@@ -192,8 +199,19 @@ if min(dimensions)==1
         savedir2=[savedir '/Post hoc/' posthoc{np}.name '/'];
         createSavedir(savedir2)
         
+        if numel(posthoc)==1
+            colorPlot=chooseColor(colorLine,1);
+        elseif numel(posthoc)==3
+            if np<3
+                colorPlot=chooseColor(colorLine,np);
+            else
+                colorPlot=[];
+            end
+        else
+            colorPlot=[];
+        end
+        
         % means
-        colorPlot=[];
         plotmean(posthoc{np}.data.continuum,CI,xlab,ylab,Fs,xlimits,nTicksX,nTicksY,colorPlot,imageFontSize,imageSize,transparancy1D,ylimits)
         legend(posthoc{np}.data.names,'Location','eastoutside','box','off')
         print('-dtiff',imageResolution,[savedir2 verifSaveName(posthoc{np}.name)])
@@ -257,7 +275,75 @@ if min(dimensions)==1
         end
     end
     
[…]
