+++ Examples/D1_ANOVA2_1rm.m
@@ -47,5 +47,5 @@ tic
 saveNplot(spmAnalysis2,...
     'savedir',savedir2,...
     'xlabel',xlab,'ylabel',ylab,'nTicksX',nTicksX,...
-    'xlimits',xlimits,'ylimits',ylimits,'nTicksY',nTicksY);
+    'xlimits',xlimits,'ylimits',ylimits,'nTicksY',nTicksY,'colorline',colorLine);
 toc
