+++ Examples/D1_ANOVA3_2rm.m
@@ -50,6 +50,6 @@ tic
 saveNplot(spmAnalysis2,...
     'savedir',savedir2,...
     'xlabel',xlab,'ylabel',ylab,'nTicksX',nTicksX,...
-    'xlimits',xlimits,'ylimits',ylimits,'nTicksY',nTicksY);
+    'xlimits',xlimits,'ylimits',ylimits,'nTicksY',nTicksY,'colorline',colorLine);
 toc
 
