### 2021-04-21 3b2b427
effectNames correction
Allow to use POST and POST2 as names for effects
