### 2021-04-22 c28f250
Update fctPostHoc1dS.m
