+++ src/saveNplot.m
@@ -195,86 +195,76 @@ posthoc=spmAnalysis.posthoc;
 dimensions=size(spmAnalysis.posthoc{1}.tTests.Tcontinuum{1});
 if min(dimensions)==1
     
-        for np=1:numel(posthoc)
-    
-            savedir2=[savedir '/Post hoc/' posthoc{np}.name '/'];
-            createSavedir(savedir2)
-    
-            if numel(posthoc)==1
-                colorPlot=chooseColor(colorLine,1);
-            elseif numel(posthoc)==3
-                if np<3
-                    colorPlot=chooseColor(colorLine,np);
-                else
-                    colorPlot=[];
-                end
+    for np=1:numel(posthoc)
+        
+        savedir2=[savedir '/Post hoc/' posthoc{np}.name '/'];
+        createSavedir(savedir2)
+        
+        %% pairewise plots
+        
+        for comp=1:numel(posthoc{np}.differences.continuum)
+            
+            % differences
+            plotmean({posthoc{np}.differences.continuum{comp}},CI,xlab,ylab,Fs,xlimits,nTicksX,[],[],imageFontSize,imageSize,transparancy1D,[])
+            if size(posthoc{np}.differences.names,1)>1
+                legend([posthoc{np}.differences.names{1,comp} ' - ' posthoc{np}.differences.names{2,comp}],'Location','eastoutside','box','off')
+                print('-dtiff',imageResolution,[savedir2 'DIFF/' verifSaveName([char(posthoc{np}.differences.names{1,comp}) ' - ' char(posthoc{np}.differences.names{2,comp})])])
+                savefig([savedir2 '/FIG/DIFF/' verifSaveName([char(posthoc{np}.differences.names{1,comp}) ' - ' char(posthoc{np}.differences.names{2,comp})])])
             else
-                colorPlot=[];
+                legend(posthoc{np}.differences.names{comp},'Location','eastoutside','box','off')
+                print('-dtiff',imageResolution,[savedir2 'DIFF/' verifSaveName(posthoc{np}.differences.names{comp})])
+                savefig([savedir2 '/FIG/DIFF/' verifSaveName(posthoc{np}.differences.names{comp})])
             end
-    
-            %% pairewise plots
-    
-            for comp=1:numel(posthoc{np}.differences.continuum)
-    
-                % differences
-                plotmean({posthoc{np}.differences.continuum{comp}},CI,xlab,ylab,Fs,xlimits,nTicksX,[],[],imageFontSize,imageSize,transparancy1D,[])
-                if min(size(posthoc{np}.differences.names{1,comp}))>1
-                    legend([posthoc{np}.differences.names{1,comp} ' - ' posthoc{np}.differences.names{2,comp}],'Location','eastoutside','box','off')
-                    print('-dtiff',imageResolution,[savedir2 'DIFF/' verifSaveName([posthoc{np}.name ' (' char(posthoc{np}.differences.names{1,comp}) ' - ' char(posthoc{np}.differences.names{2,comp}) ')'])])
-                    savefig([savedir2 '/FIG/DIFF/' verifSaveName([posthoc{np}.name ' (' char(posthoc{np}.differences.names{1,comp}) ' - ' char(posthoc{np}.differences.names{2,comp}) ')'])])
-                else
-                    legend(posthoc{np}.differences.names{comp},'Location','eastoutside','box','off')
-                    print('-dtiff',imageResolution,[savedir2 'DIFF/' verifSaveName(posthoc{np}.differences.names{comp})])
[…]
