+++ src/fctSPM.m
@@ -36,20 +36,20 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 % OUTPUTS
-%`spmAnalaysis.mat` is a structure composed of the results of teh statistical analysis (ANOVA + Post Hoc). 
+%`spmAnalaysis.mat` is a structure composed of the results of teh statistical analysis (ANOVA + Post Hoc).
 
-%`spmAnalysis.anova` is composed of different fields :  
+%`spmAnalysis.anova` is composed of different fields :
 %* `type` is the type of ANOVA performed
 %* `effectNames` is a structure (one cell for each effect) that represent the names of the effects tested (mains and interactions)
-%* `alpha` is the alpha risk choosen for the anova (default is 0.05 (5%)).  
+%* `alpha` is the alpha risk choosen for the anova (default is 0.05 (5%)).
 %* `pCritical` is the alpha risk used for the anova. Warning message is displayed if this value is modified.
 %* `nIterations` is the number of iterations performed for the anova.
 %* `maxIterations` is the number of maximal iterations possible for the anova.
 %* `Fcontinuum` is a structure that represent the F-value for each node
 %* `Fthreshold` is a structure that represent the statistical threshold for the F-values (statistical inference)
-%* `Fsignificant` is a structure that contains the logical for the significance (Fcontinuum > Fthreshold) of each effect of the ANOVA (1 if significant, 0 if not).  
+%* `Fsignificant` is a structure that contains the logical for the significance (Fcontinuum > Fthreshold) of each effect of the ANOVA (1 if significant, 0 if not).
 %* `clusterLocation` is a structure (one for each significant cluster) that contains the location (start and end as indexes) of each significant cluster.
-%* `clusterP` is a structure (one for each significant cluster) that contains the p-value of each significant cluster. 
+%* `clusterP` is a structure (one for each significant cluster) that contains the p-value of each significant cluster.
 
 %`clusterLocation` and `clusterP` are created only in one dimension
 
@@ -81,7 +81,7 @@
 %`clusterLocation` and `clusterP` are created only in one dimension and are not corrected with the result of the ANOVA.
 
 %* `tTests.contourSignificant` represents a modified T-value continuum to display smoother contour plots.
-%`tTests.contourSignificant is created only in two dimensions.  
+%`tTests.contourSignificant is created only in two dimensions.
 
 
 
@@ -89,21 +89,21 @@
 %Two folders composed of figures in `.TIF` and `.fig` are created
 
 %#### ANOVA ####
-%Each effect is also display on a specific figure in `.TIF` format named after `effectNames`, a floder named FIG is also created and contains the figures in `.fig` format.   
+%Each effect is also display on a specific figure in `.TIF` format named after `effectNames`, a floder named FIG is also created and contains the figures in `.fig` format.
 %For 1D:  The curve represents the `Fcontinuum`, the horizontal line is the `Fthreshold`, the highlighted parts of the curve in blue represents the significant cluster(s), and the vertical lines are the start and end indexes.
 %For 2D: The map represents the `Fcontinuum`, with a colorbar maximum at `Fthreshold`, the white clusters represent the significant clusters.
 
 %#### Post Hoc ####
-%This folder contains additional folders (0 (for anova1), 3 (for anova2) or 7 (for anova3)) that contain figures and metrics representing the different post-hoc tests.  
+%This folder contains additional folders (0 (for anova1), 3 (for anova2) or 7 (for anova3)) that contain figures and metrics representing the different post-hoc tests.
 
[…]
