+++ src/plot/PlotmeanSub.m
@@ -1,4 +1,4 @@
-function PlotmeanSub(mapsAll,nameSub,effectsRm,effectNames,savedir,xlab,ylab,Fs,imageResolution,CI,ylimits,nTicksX,nTicksY,xlimits,imageFontSize,imageSize,colorLine,colorMap,colorbarLabel,limitMeanMaps)
+function PlotmeanSub(mapsAll,nameSub,effectsRm,effectNames,savedir,xlab,ylab,Fs,imageResolution,CI,ylimits,nTicksX,nTicksY,xlimits,imageFontSize,imageSize,colorLine,colorMap,colorbarLabel,limitMeanMaps,transparancy1D)
 nSuj=size(mapsAll,1);
 if isempty(nameSub)
     for i=1:nSuj
@@ -10,7 +10,7 @@ for i=1:nSuj
     for cond=1:size(effectsRm{1},2)
         if min(size(mapsAll{i,cond}))==1
             Dataplot{1}=mapsAll{i,cond}';
-            plotmean(Dataplot,CI,xlab,ylab,Fs,xlimits,nTicksX,colorLine,imageFontSize,imageSize)
+            plotmean(Dataplot,CI,xlab,ylab,Fs,xlimits,nTicksX,nTicksY,colorLine,imageFontSize,imageSize,transparancy1D,ylimits)
         else
             Dataplot=mapsAll{i,cond};
             displayMeanMaps(Dataplot,Fs,xlab,ylab,ylimits,nTicksX,nTicksY,limitMeanMaps,xlimits,imageFontSize,imageSize,colorbarLabel,colorMap)
