### 2021-05-07 f6b54e6
lineStyle + noMeans
add the possibility to choose the lineStyle of the curves

add the possibility to display only the mean without the standard deviation
