+++ src/utility/chooseColor.m
@@ -1,17 +1,0 @@
-function [colorPlot]=chooseColor(colorLine,effectN)
-
-
-if isempty(colorLine)
-    colorPlot=[];
-elseif iscell(colorLine)
-    if numel(colorLine)>=effectN
-        colorPlot=colorLine{effectN};
-    else
-        colorPlot=[];
-    end
-else
-    colorPlot=colorLine;
-end
-
-
-end
