### 2021-05-07 c99bd69
input modifications
CI description was modified
