### 2021-05-17 e368cf5
bug fix
plotmeans CI and SEM
