### 2021-05-17 e8725f0
warnings/figure display
desactivate and reactivate figure display
activate warnings
