### 2021-06-08 tramarobin/fctSnPM#19 @nicholst
JOSS review questions
Hi,

I'm providing a much belated review for JOSS.  Please consider the following issues...
<hr/>

[README.md](https://github.com/tramarobin/fctSPM/blob/master/README.md):
 > Post-hoc tests with Bonferonni correction are only approximate.

Can you clarify if this is the usual "Bonferonni is conservative", or something more fundamental relating to a post-hoc test being viewed only after a significant main effect.  Not asking for a huge exposition, but just e.g. "... Bonferonni correction are conservative" or  "... Bonferonni correction are only approximate because..."

**Typos/spelling errors**: please run a spell check on the documentation (e.g. "This function synthetises" ->"This function synthesizes"; or "results of teh statistical analysis")

`.\fctSPM\Examples\D1_ANOVA2_1rm.m` is a windows-centric path specification and Matlab doesn't support executing a script with a path.  Presumably it would be better to say "The output of `D1_ANOVA2_1rm.m` in the Examples folder is as follows" (or something).

**Input types**:  I must say I like the clear designation of input type in the documentation of inputs, e.g.

> `savedir` is the path ... might be overwritten @ischar."

but it is done inconsistently.  I.e. the main inputs don't have these end-of-sentence tags (and `IT` doesn't have any), and they are typset inconsistently when given (sometimes preceeded by ".", "..." or just space).  Maybe just integrate them into the prose description or consistently add them in parentheses at the end (`char`).

**iterations**:  As permutation tests are used here, I kept search the documentation and code:  How many permutations are used?   Eventually I figured out that (please correct me if not!) that "iterations" is supposed to be read as "permutations".   Iterations is usually understood in the context of optimisation, iteratively improving an objective function.  I can't expect you to replace a term that appears in 132 lines of code, but please, at least at the first reference of "iteration" in the documentation please clarify this is the number of permutations.

<hr/>

[fctSPM.m](https://github.com/tramarobin/fctSPM/blob/master/src/fctSPM.m) & [fctSPMS.m](https://github.com/tramarobin/fctSPM/blob/master/src/fctSPMS.m)

"Information" is always singular, but multiple times it is used as "Informations"; also, the lines starting "Please read README.md" have a strange nonprinting character before "Please".

Also very minor, the hypen in the Nichols & Holmes reference isn't plain ASCII; its better to replace "" with vanilla "-".
