### 2021-06-08 70327cd
Change defaults in examples so they run without edits
