### 2021-06-08 #3159 @nicholst
This is a nicely documented and focused statistical Matlab tool that provides inference for ANOVA designs of 1D and 2D functional data by extending the spm1d package.  While the review guidelines of JOSS specified that tests should be examined and verified, I didn't find any tests in the repo; I reviewed the documentation and examples.

I have raised a few superficial comments in this issue: https://github.com/tramarobin/fctSPM/issues/19

and made some suggests to make the Examples easier to use (i.e. so they are ready to run as-is) here: https://github.com/tramarobin/fctSPM/pull/20
