### 2021-06-09 9fd3537
Permutation modifications
Iterations and according variables were modified as Permutations
