+++ Article@JOSS/paper.md
@@ -33,7 +33,7 @@ To analyze all the dimensions of a signal without losing information, the analys
 # fctSPM
 The function we propose meets two objectives. 1/ to allow statistical inferences on curves and maps with a standardized format and 2/ to simplify analyses by comparing means while considering intersections with tests performed upstream (ANOVA and post-hoc of main effects).
 
-The statistical tests are performed taking into account the independent and repeated measure effects provided in the obligatory function inputs. ANOVA, up to three-way ANOVA with three repeated measures, is performed if required, and followed by post-hoc tests as paired or independent Student t-tests. By default, the ANOVA is performed with an alpha risk of 5%, while post-hoc tests alpha risk is adjusted with Bonferronni correction. A number of 10/alpha iterations (200 for a 5% risk) is defined for each test. Statistical parameters are customizable via optional inputs, like `multiIT` which can be used to increase the number of iterations and achieve numerical stability and reliable analysis [@Nichols:2002]. A Matlab (.mat) file containing the number of permutations, the significant clusters, the statistical thresholds, and the raw data used in the analysis is also generated for each test family.
+The statistical tests are performed taking into account the independent and repeated measure effects provided in the obligatory function inputs. ANOVA, up to three-way ANOVA with three repeated measures, is performed if required, and followed by post-hoc tests as paired or independent Student t-tests. By default, the ANOVA is performed with an alpha risk of 5%, while post-hoc tests alpha risk is adjusted with Bonferronni correction. A number of 10/alpha permutations (200 for a 5% risk) is defined for each test. Statistical parameters are customizable via optional inputs, like `multiPerm` which can be used to increase the number of permutations and achieve numerical stability and reliable analysis [@Nichols:2002]. A Matlab (.mat) file containing the number of permutations, the significant clusters, the statistical thresholds, and the raw data used in the analysis is also generated for each test family.
 
 To interpret the results, figures directly usable for presentations and/or articles are available. In one dimension, the main figure contains the mean and standard deviations for each group of the analyzed condition, and the results of the post-hoc tests corrected with the result of the ANOVA. In two dimensions, the mean maps and the standard deviations are on two separate figures, and the result of the statistical analysis is displayed on the map of differences between two modalities. Therewith, other figures that display absolute and relative differences, effect sizes, and the raw value of the statistical test and its threshold are available.
 
