+++ src/fctPostHoc1d.m
@@ -1,4 +1,4 @@
-function posthoc=fctPostHoc1d(nEffects,indicesEffects,maps1d,dimensions,modalitiesAll,typeEffectsAll,eNames,savedir,multiIterations,IT,xlab,ylab,Fs,imageResolution,IC,ylimits,nx,ny,xlimits,anovaEffects,maximalIT,colorLine,doAllInteractions,imageFontSize,imageSize,alphaT,alphaAOV,transparancy1D,ratioSPM,yLimitES,spmPos,aovColor,lineStyle)
+function posthoc=fctPostHoc1d(nEffects,indicesEffects,maps1d,dimensions,modalitiesAll,typeEffectsAll,eNames,savedir,multiPerm,Perm,xlab,ylab,Fs,imageResolution,IC,ylimits,nx,ny,xlimits,anovaEffects,maximalPerm,colorLine,doAllInteractions,imageFontSize,imageSize,alphaT,alphaAOV,transparancy1D,ratioSPM,yLimitES,spmPos,aovColor,lineStyle)
 close all
 
 savedir=[savedir '/Post hoc/'];
@@ -101,16 +101,16 @@ if nEffects==1
             
             % inference
             posthoc{1}.tTests.names=posthoc{1}.differences.names;
-            [posthoc{1}.tTests.nWarning{comp},iterations,alpha]=fctWarningIterations(Ttest,pBonferroni,multiIterations,maximalIT,IT);
+            [posthoc{1}.tTests.nWarning{comp},permutations,alpha]=fctWarningPermutations(Ttest,pBonferroni,multiPerm,maximalPerm,Perm);
             posthoc{1}.tTests.alpha=alphaT;
             if warnT==1
                 posthoc{1}.tTests.warning="alpha is not valid";
             end
             posthoc{1}.tTests.pBonferroni=pBonferroni;
             posthoc{1}.tTests.pCritical{comp}=alpha;
-            posthoc{1}.tTests.nIterations{comp}=iterations;
-            Ttest_inf=Ttest.inference(alpha,'iterations',iterations,'force_iterations',logical(1),'two_tailed',logical(1));
-            posthoc{1}.tTests.maxIterations{comp}=Ttest_inf.nPermUnique;
+            posthoc{1}.tTests.nPermutations{comp}=permutations;
+            Ttest_inf=Ttest.inference(alpha,'permutations',permutations,'force_iterations',logical(1),'two_tailed',logical(1));
+            posthoc{1}.tTests.maxPermutations{comp}=Ttest_inf.nPermUnique;
             posthoc{1}.tTests.Tthreshold{comp}=Ttest_inf.zstar;
             posthoc{1}.tTests.Tsignificant{1,comp}=zeros(dimensions(1),dimensions(2));
             posthoc{1}.tTests.Tcontinuum{1,comp}=reshape(Ttest_inf.z,dimensions(1),dimensions(2));
@@ -282,16 +282,16 @@ if nEffects==2
                 
                 % inference
                 posthoc{mainEffect(1)}.tTests.names=posthoc{mainEffect(1)}.differences.names;
-                [posthoc{mainEffect(1)}.tTests.nWarning{comp},iterations,alpha]=fctWarningIterations(Ttest,pBonferroni,multiIterations,maximalIT,IT);
+                [posthoc{mainEffect(1)}.tTests.nWarning{comp},permutations,alpha]=fctWarningPermutations(Ttest,pBonferroni,multiPerm,maximalPerm,Perm);
                 posthoc{mainEffect(1)}.tTests.alpha=alphaT;
                 if warnT==1
                     posthoc{mainEffect(1)}.tTests.warning="alpha is not valid";
                 end
[…]
