+++ src/utility/openfigINV.m
@@ -1,0 +1,5 @@
+function openfigINV
+
+openfig(uigetfile,'visible')
+
+end

[…]
