### 2021-06-09 0f66779
Permutations modifications
