+++ README.md
@@ -46,7 +46,7 @@ Analyzing continuous values (i.e., time series) can provide more information tha
 In addition, oscillatory signals such as muscle vibrations and electromyograms contain information in the temporal and frequency domains. 
 Once again, scalar analysis reduces the information at only one dimension by discarding two dimensions among the magnitude and the localization in the time and/or frequency domain.
 
-Factorial SPM (fctSPM) allows MATLAB users to create figures of 1D and 2D [SPM](https://spm1d.org/) analysis for ANOVA and post-hoc designs.
+Factorial SPM (fctSPM) allows MATLAB users to create figures of 1D and [2D SPM](https://spm1d.org/doc/Stats2D/ex2d_matlab.html) analysis for ANOVA and post-hoc designs.
 The statistical analysis is also saved in `.mat` files.   
 This function synthetizes the main and interaction effects to display only the significant post-hoc regarding the results of the ANOVA.   
 For post-hoc for interaction effects, the main effect is also displayed if located elsewhere than the interaction effect.
@@ -54,8 +54,8 @@ For post-hoc for interaction effects, the main effect is also displayed if locat
 
 ## Citing fctSPM ##
 * This function: under review in JOSS   
-* for spm1d: Pataky TC (2010). Generalized n-dimensional biomechanical field analysis using statistical parametric mapping. Journal of Biomechanics 43, 1976-1982.   
-* for permutation tests: Nichols TE, Holmes AP (2002). Nonparametric permutation tests for functional neuroimaging: a primer with examples. Human Brain Mapping 15(1), 1–25.   
+* For spm1d: Pataky TC (2010). Generalized n-dimensional biomechanical field analysis using statistical parametric mapping. Journal of Biomechanics 43, 1976-1982.   
+* For permutation tests: Nichols TE, Holmes AP (2002). Nonparametric permutation tests for functional neuroimaging: a primer with examples. Human Brain Mapping 15(1), 1–25.   
 * spm1d package for matlab can be downloaded at: https://github.com/0todd0000/spm1dmatlab   
 
 
@@ -196,15 +196,9 @@ spmAnalysis=fctSPM(data, independantEffects, repeatedMeasuresEffects, varargin)
 ```
 
 ### Obligatory inputs ###
-* `data` is a x by y cell array.  
-x corresponds to a subjects and y corresponds to a repeated measure.     
-Each cell contains a column vector (1D) or a matrix (2D) corresponding to the mean of the subject and condition.
-* `independantEffects` is a cell array defining the independent effects.  
-1 cell by effect.  
-Each cell contains the name ('char') of each modality for the given subject and must correspond to the number of subjects/rows (x).
-* `repeatedMeasuresEffects` is a cell array defining the repeated measure effects.  
-1 cell by effect.    
-Each cell contains the name ('char') of each modality for the given condition and must correspond to the number of conditions/columns (y). 
+* `data` is a x by y cell array, with x corresponding to subjects and y corresponding to repeated measures. Each cell contains a column vector (1D) or a matrix (2D) corresponding to the mean of the subject and condition.
+* `independantEffects` is a cell array defining the independent effects. There must be 1 cell by effect. Each cell contains the name ('char') of each modality for the given subject and must correspond to the number of subjects/rows (x). For instance `independantEffects{1}={'Male' 'Male' 'Female' 'Female'}` if the two first subjects are `Male` and the two last are `Female`.
[…]
