### 2021-06-09 tramarobin/fctSPM#19 @nicholst
Thanks for the quick response @tramarobin.

Thanks for the updates.  It all looks good... all that's left is trivial:

You corrected the Nichols & Holmes 'dash' in the 2nd but not the first instance in fctSPM*.m (look after "please cite for permutation").

Also, totally trivial, but might as well fix it:  Mostly you have references to Unix-style paths and "/" as path separator.    But I just noticed ...
```
fctSPM.m:% Examples are in ...\fctSPM\Examples
fctSPM.m:% see .\fctSPM\Examples for help
fctSPMS.m:% Examples are in ...\fctSPM\Examples
fctSPMS.m:% see .\fctSPM\Examples for help
saveNplot.m:% see .\fctSPM\Examples for help
```

I would simply avoid this by "Examples are in the Examples directory" ... there can be no confusion about that I think.

OTW, good to go!
