### 2021-06-09 tramarobin/fctSPM#19 @tramarobin
> You corrected the Nichols & Holmes 'dash' in the 2nd but not the first instance in fctSPM*.m (look after "please cite for permutation").

Oh sorry I forgot this one. It should be ok now.

> I would simply avoid this by "Examples are in the Examples directory" ... there can be no confusion about that I think.

I modified this line as follow : `You can find different scripts creating output for 1D and 2D data in ...\fctSPM\Examples`
