### 2021-06-09 tramarobin/fctSPM#21 @tramarobin
Function's name
@nicholst & @0todd0000, thank you for your review and insights on this function.

I was just wondering if fctSnPM, for Statistical nonParametric Mappping, would not be more appropriate name for this function as the approach use to perform the statistical inferences is non-parametric.

If you are ok with this change, I'll modify the different files to be with the same name.
