### 2021-06-09 tramarobin/fctSPM#19 @nicholst
Ok… but my other point was that, to my Linux eyes, I stumble over a path separator of “\” and I saw “/“ elsewhere somewhere in the code (I think), and so suggested a formulation that avoided path separators altogether…. But honestly I’m not that fussed…
