### 2021-06-09 tramarobin/fctSPM#21 @nicholst
Im biased, but, yes I think this is an excellent proposal. :)
