### 2021-06-09 tramarobin/fctSPM#21 @0todd0000
I agree, good suggestion!
