+++ src/fctSPM.m
@@ -13,7 +13,7 @@
 % The type of anova (if required) and post-hoc are choosen regarding the independant or repeated measure effect given in parameters.
 % The function automatically adapts to 1D and 2D data
 % Analysis and figures of the analysis are saved
-% Examples are in ...\fctSPM\Examples
+% You can find different scripts creating output for 1D and 2D data in ...\fctSPM\Examples
 % 1D examples are torque ratios
 % 2D examples are maps obtained with continuous wavelet transforms
 
@@ -21,7 +21,7 @@
 % spm1d package for matlab is published elswhere : https://github.com/0todd0000/spm1dmatlab
 
 % please cite for spm1d : Pataky TC (2010). Generalized n-dimensional biomechanical field analysis using statistical parametric mapping. Journal of Biomechanics 43, 1976-1982.
-% please cite for permutation tests : Nichols TE, Holmes AP (2002). Nonparametric permutation tests for functional neuroimaging: a primer with examples. Human Brain Mapping 15(1), 1–25.
+% please cite for permutation tests : Nichols TE, Holmes AP (2002). Nonparametric permutation tests for functional neuroimaging: a primer with examples. Human Brain Mapping 15(1), 1-25.
 
 % WARNINGS:
 
@@ -136,7 +136,7 @@
 % also avoid underscore (_) or minus (-) sign. Spaces are OK
 
 % OPTIONAL
-% see the description at begining of the function (inputParser)
+% see the description at begining of the function (inputParser) or on GitHub (https://github.com/tramarobin/fctSPM#optional-inputs)
 % see .\fctSPM\Examples for help
 
 %% Information
