+++ src/fctSPMS.m
@@ -96,7 +96,7 @@
 
 % OPTIONAL
 % see the description at begining of the function (inputParser) or on GitHub (https://github.com/tramarobin/fctSPM#optional-inputs)
-% see .\fctSPM\Examples for help
+% see ./fctSPM/Examples for help
 
 %% Information
 % All the dataset must be balanced for ANOVA 2 and 3
