### 2021-06-10 tramarobin/fctSPM#19 @tramarobin
Ok, sorry, I did not understood your remark on this point, but I do now.

If it works on Linux too (I only tested on Windows and Mac), It would be better to let it as it is I think.
I did not now the `fullfile` function until now, and although it would be a better formulation, it would add nothing if it already works fine (and also save me a lot of time as many files are created with different paths...)

I checked the different lines of code and change some '\' in '/'. I hope it wont be any left.
