+++ src/fctSPM.m
@@ -156,7 +156,7 @@ function spmAnalysis=fctSPM(mapsAll,effectsInd,effectsRm,varargin)
 p = inputParser;
 
 % utilities
-addParameter(p,'savedir',[]); % path to save directory
+addParameter(p,'savedir',''); % path to save directory
 addParameter(p,'effectsNames',{'A','B','C'},@iscell); % name of the different effect tested (changes the name of folder and files)
 % the independant effects must be named first
 addParameter(p,'plotSub',0,@isnumeric) % 1 to plot the mean of each subject
