+++ src/saveNplot.m
@@ -52,7 +52,7 @@
 
 % OPTIONAL
 % see the description at begining of the function (inputParser) or on GitHub (https://github.com/tramarobin/fctSPM#optional-inputs)
-% see .\fctSPM\Examples for help
+% see ./fctSPM/Examples for help
 
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
