+++ Examples/D1_ANOVA1rm.m
@@ -1,11 +1,16 @@
 % This function takes ~35 seconds on i5 processor and files takes 5 Mo of
 % storage
-% don't forget to add the source code path
 
 clear
 close all
 clc
 
+% Don't forget to add the source code path
+% This funtion will automatically create multiple files in a save directory located at the savedir
+% adress
+
+savedir=''; % is the adress where the output of fctSnPM is saved 
+savedir2=savedir; % is the adress where the output of saveNplot is saved 
 
 % data
 load ExampleDatas
@@ -19,8 +24,6 @@ colorLine{1}=[rgb('green'); rgb('blue'); rgb('red'); rgb('black')];
 independantEffects=[];
 % There are 20 subjects
 
-savedir=[];
-savedir2=[];
 xlab='Angle (°)';
 ylab='Ratio';
 xlimits=[30 90];
