+++ Examples/D2_ANOVA2rm.m
@@ -1,10 +1,16 @@
 % This function takes ~60 seconds on i5 processor and files takes 30 Mo of
 % storage
-% don't forget to add the source code path
 
 clear
 close all
 clc
+
+% Don't forget to add the source code path
+% This funtion will automatically create multiple files in a save directory located at the savedir
+% adress
+
+savedir=''; % is the adress where the output of fctSnPM is saved 
+savedir2=savedir; % is the adress where the output of saveNplot is saved 
 
 
 % data
@@ -22,8 +28,7 @@ effectNames={'Device','Activation'};
 % Data(:,3) correspond to Device=US and Activation=Contracted
 % Data(:,4) correspond to Device=US and Activation=Relaxed
 
-savedir=[];
-savedir2=[];
+
 xlab='Time (s)';
 ylab='Frequency (Hz)';
 Fs=500;
