+++ Examples/D1_pairedTtest.m
@@ -1,12 +1,16 @@
 % This function takes ~15 seconds on i5 processor and files takes 2 Mo of
 % storage
-% don't forget to add the source code path
-
 
 clear
 close all
 clc
 
+% Don't forget to add the source code path
+% This funtion will automatically create multiple files in a save directory located at the savedir
+% adress
+
+savedir=''; % is the adress where the output of fctSnPM is saved 
+savedir2=savedir; % is the adress where the output of saveNplot is saved 
 
 % data
 load ExampleDatas
@@ -18,8 +22,6 @@ independantEffects=[]; % empty
 repeatedMeasuresEffects{1}={'Right','Left'}; % Side
 
 
-savedir=[];
-savedir2=[];
 xlab='Angle (°)';
 ylab='Ratio';
 xlimits=[30 90];
