+++ Examples/D2_ANOVA3_2rm.m
@@ -1,9 +1,15 @@
 % This function takes ~250 seconds on i5 processor and files takes 170 Mo
-% don't forget to add the source code path
 
 clear
 close all
 clc
+
+% Don't forget to add the source code path
+% This funtion will automatically create multiple files in a save directory located at the savedir
+% adress
+
+savedir=''; % is the adress where the output of fctSnPM is saved 
+savedir2=savedir; % is the adress where the output of saveNplot is saved 
 
 
 % data
@@ -23,8 +29,7 @@ effectNames={'Group','Device','Activation'};
 % Data(:,3) correspond to Device=US and Activation=Contracted
 % Data(:,4) correspond to Device=US and Activation=Relaxed
 
-savedir='C:\Users\LIBM_yb\Documents\DATA_MATLAB\spmTest\1';
-savedir2='C:\Users\LIBM_yb\Documents\DATA_MATLAB\spmTest\2';
+
 xlab='Time (s)';
 ylab='Frequency (Hz)';
 Fs=500;
