+++ Examples/D1_independantTtest.m
@@ -1,10 +1,16 @@
 % This function takes ~15 seconds on i5 processor and files takes 1 Mo of
 % storage
-% don't forget to add the source code path
 
 clear
 close all
 clc
+
+% Don't forget to add the source code path
+% This funtion will automatically create multiple files in a save directory located at the savedir
+% adress
+
+savedir=''; % is the adress where the output of fctSnPM is saved 
+savedir2=savedir; % is the adress where the output of saveNplot is saved 
 
 % data
 load ExampleDatas
@@ -16,8 +22,7 @@ linestyle{1}={'--' ':'};
 independantEffects{1}={'M','M','M','F','M','F','F','M','F','M','M','F','F','M','F','F','M','F','F','M'}; % same number than participants
 repeatedMeasuresEffects=[]; % empty
 
-savedir=[];
-savedir2=[];
+
 xlab='Angle (°)';
 ylab='Ratio';
 xlimits=[30 90];

[…]
