### 2021-06-10 #3159 @tramarobin
@cMadan,  after a quick discussion with the reviewers (https://github.com/tramarobin/fctSPM/issues/21), I would like to know if it was possible to change the name of the function, as well as the name of the paper.

The function's name would be `fctSnPM`, and the title of the article `'fctSnPM: Factorial ANOVA and post-hoc tests for Statistical nonParametric Mapping in MATLAB'`

This name better suits what the function does as it employs a non parametric approach.

Let me know if it is possible, and in that case I'll made the necessary changes.
