### 2021-06-10 ff80478
Example modifications
Add the funtion onlyPlot to plot only a part of the analysis without saving the output.
