+++ src/onlyPlot.m
@@ -1,0 +1,335 @@
+% Trama Robin (LIBM) 06/2021 --> JOSS
+% trama.robin@gmail.com
+
+% available at :
+% - https://github.com/tramarobin/fctSPM
+% - https://www.mathworks.com/matlabcentral/fileexchange/77945-fctspm
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+% plot a part of the analysis obtain with `fctSPM` and `fctSPMS`.
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+%% INPUTS
+% OBLIGATORY
+% spmAnalysis obtained with `fctSPM` or `fctSPMS`.
+
+% OPTIONAL
+% see the description at begining of the function (inputParser) or on GitHub (https://github.com/tramarobin/fctSPM#optional-inputs)
+% see ./fctSPM/Examples for help
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+function onlyPlot(spmAnalysis,varargin)
+
+%% Optional inputs
+p = inputParser;
+
+% general plot parameters
+addParameter(p,'ylabel','',@ischar); % label of Y-axis
+addParameter(p,'xlabel','',@ischar); % label of X-axis
+addParameter(p,'samplefrequency',1,@isnumeric); % change xticks to correspond at the specified frequency
+addParameter(p,'xlimits',[],@isnumeric); % change xticks to correspond to the specified range (can be negative)
+% specified either samplefrequency or xlimits, but not both
+addParameter(p,'nTicksX',5,@isnumeric); % number of xticks displayed
+addParameter(p,'nTicksY',[],@isnumeric); % number of yticks displayed
+addParameter(p,'imageSize',[],@isnumeric) % size of the image in cm. X --> X*X images, [X Y] X*Y imgages. By default the unit is pixels [0 0 720 480].
+addParameter(p,'imageFontSize',12,@isnumeric) % font size of images
+addParameter(p,'ylimits',[],@isnumeric); % change yticks to correspond to the specified range
+addParameter(p,'linestyle','-') % In 1D : lineStyle for plots (default  is continuous) // Specify linestyle for each modality in cell, apply each style to each modality (independant effect first) // In 2D :linestyle of the contour plot
+
+% 2d plot parameters
+addParameter(p,'colorMap',cbrewer('seq','Reds', 64)) % colormap used for means and ANOVA and ES plots (0 to positive)
+addParameter(p,'colorMapDiff',flipud(cbrewer('div','RdBu', 64))) % colormap used for differences and SPM plot (0 centered)
+addParameter(p,'colorbarLabel','',@ischar); % name of the colorbar label
+addParameter(p,'limitMeanMaps',[],@isnumeric); % limit of the colorbar. the value of X will make the colorbar going from 0 to X for all plots (easier to compare). If not specified, the maps wont necessery be with the same range but will be automatically scaled
+addParameter(p,'displaycontour',1,@isnumeric); % display contour map on differences and size effect maps (0 to not display)
+addParameter(p,'contourcolor','w'); % color of the line for the contour plot
+addParameter(p,'dashedColor',[0 0 0],@isnumeric) % color of the dashed zone of the contour plot (default is white)
+addParameter(p,'transparancy',50,@isnumeric) % transparancy of the dashed zone (0=transparant, 255=opaque)
[…]
