+++ Examples/D1_ANOVA2_1rm.m
@@ -45,14 +45,20 @@ spmAnalysis=fctSPM(DATA,independantEffects,repeatedMeasuresEffects,...
     'xlabel',xlab,'ylabel',ylab,'nTicksX',nTicksX,...
     'xlimits',xlimits,'ylimits',ylimits,'nTicksY',nTicksY,'colorline',colorLine,'lineStyle',lineStyle);
 toc
-% 
-% tic
-% spmAnalysis2=fctSPMS(DATA,independantEffects,repeatedMeasuresEffects,'effectsNames',effectNames);
-% toc
-% 
-% tic
-% saveNplot(spmAnalysis2,...
-%     'savedir',savedir2,...
-%     'xlabel',xlab,'ylabel',ylab,'nTicksX',nTicksX,...
-%     'xlimits',xlimits,'ylimits',ylimits,'nTicksY',nTicksY,'colorline',colorLine);
-% toc
+
+tic
+spmAnalysis2=fctSPMS(DATA,independantEffects,repeatedMeasuresEffects,'effectsNames',effectNames);
+toc
+
+tic
+onlyPlot(spmAnalysis2,...
+    'xlabel',xlab,'ylabel',ylab,'nTicksX',nTicksX,...
+    'xlimits',xlimits,'ylimits',ylimits,'nTicksY',nTicksY,'colorline',colorLine);
+toc
+
+tic
+saveNplot(spmAnalysis2,...
+    'savedir',savedir2,...
+    'xlabel',xlab,'ylabel',ylab,'nTicksX',nTicksX,...
+    'xlimits',xlimits,'ylimits',ylimits,'nTicksY',nTicksY,'colorline',colorLine);
+toc
