+++ Examples/D2_ANOVA3_2rm.m
@@ -55,6 +55,14 @@ spmAnalysis2=fctSPMS(DATA,independantEffects,repeatedMeasuresEffects,'effectsNam
 toc
 
 tic
+onlyPlot(spmAnalysis2,...
+    'xlabel',xlab,'ylabel',ylab,...
+    'sampleFrequency',Fs,'colorbarLabel',colorbarLabel,...
+    'limitMeanMaps',limitMeanMaps,...
+    'ylimits',ylimits,'nTicksX',nTicksX,'nTicksY',nTicksY);
+toc
+
+tic
 saveNplot(spmAnalysis2,...
     'savedir',savedir2,...
     'xlabel',xlab,'ylabel',ylab,...
