+++ Examples/D1_independantTtest.m
@@ -44,6 +44,12 @@ spmAnalysis2=fctSPMS(DATA,independantEffects,repeatedMeasuresEffects,'effectsNam
 toc
 
 tic
+onlyPlot(spmAnalysis2,...
+    'xlabel',xlab,'ylabel',ylab,'nTicksX',nTicksX,...
+    'xlimits',xlimits,'ylimits',ylimits);
+toc
+
+tic
 saveNplot(spmAnalysis2,...
     'savedir',savedir2,...
     'xlabel',xlab,'ylabel',ylab,'nTicksX',nTicksX,...
