### 2021-06-14 32da215
fctSnPM
Change the name of the files, documentation and article to correspond to fctSnPM instead of fctSPM
