+++ Examples/D2_ANOVA1.m
@@ -9,7 +9,7 @@ clc
 % This funtion will automatically create multiple files in a save directory located at the savedir
 % adress
 
-savedir=''; % is the adress where the output of fctSPM is saved 
+savedir=''; % is the adress where the output of fctSnPM is saved 
 savedir2=savedir; % is the adress where the output of saveNplot is saved 
 
 
@@ -36,9 +36,9 @@ nTicksY=7;
 nTicksX=6;
 
 
-%SPM
+%SnPM
 tic
-spmAnalysis=fctSPM(DATA,independantEffects,repeatedMeasuresEffects,...
+snpmAnalysis=fctSnPM(DATA,independantEffects,repeatedMeasuresEffects,...
     'savedir',savedir,...
     'effectsNames',effectNames,...
     'xlabel',xlab,'ylabel',ylab,...
@@ -48,11 +48,11 @@ spmAnalysis=fctSPM(DATA,independantEffects,repeatedMeasuresEffects,...
 toc
 
 tic
-spmAnalysis2=fctSPMS(DATA,independantEffects,repeatedMeasuresEffects,'effectsNames',effectNames);
+snpmAnalysis2=fctSnPMS(DATA,independantEffects,repeatedMeasuresEffects,'effectsNames',effectNames);
 toc
 
 tic
-onlyPlot(spmAnalysis2,...
+onlyPlot(snpmAnalysis2,...
     'xlabel',xlab,'ylabel',ylab,...
     'sampleFrequency',Fs,'colorbarLabel',colorbarLabel,...
     'limitMeanMaps',limitMeanMaps,...
@@ -60,7 +60,7 @@ onlyPlot(spmAnalysis2,...
 toc
 
 tic
-saveNplot(spmAnalysis2,...
+saveNplot(snpmAnalysis2,...
     'savedir',savedir2,...
     'xlabel',xlab,'ylabel',ylab,...
     'sampleFrequency',Fs,'colorbarLabel',colorbarLabel,...
