+++ src/utility/createSavedir2d.m
@@ -4,10 +4,10 @@ warning('off', 'MATLAB:MKDIR:DirectoryExists');
 
 mkdir([savedir '/DIFF'])
 mkdir([savedir '/ES'])
-mkdir([savedir '/SPM'])
+mkdir([savedir '/SnPM'])
 mkdir([savedir '/SD'])
 mkdir([savedir '/FIG/DIFF'])
 mkdir([savedir '/FIG/ES'])
-mkdir([savedir '/FIG/SPM'])
+mkdir([savedir '/FIG/SnPM'])
 mkdir([savedir '/FIG/SD'])
 end
