### 2021-06-14 ccf37c5
fctSnPM
minor corrections for fctSnPM
