+++ Examples/D2_ANOVA2rm.m
@@ -9,8 +9,8 @@ clc
 % This funtion will automatically create multiple files in a save directory located at the savedir
 % adress
 
-savedir=''; % is the adress where the output of fctSnPM is saved 
-savedir2=savedir; % is the adress where the output of saveNplot is saved 
+savedir=''; % is the adress where the output of fctSnPM is saved
+savedir2=savedir; % is the adress where the output of saveNplot is saved
 
 
 % data
