### 2021-06-22 #3159 @nicholst
Hi @cMadan - I'm actually struggling with something simple.  The guidelines mention 'the paper'  ("Does the full list of paper authors seem appropriate and complete?") but I can't find 'the paper'... where is it?
