### 2021-07-14 #3159 @cMadan
@tramarobin, everything looks good to me! We could be finalised within a day or two from this point.

To move forward with accepting your submission, there are a few last things to take care of:

- [ ] Make a tagged release of your software (that now includes the revisions related to the JOSS review), and list the version tag of the archived version here.

You may find this helpful: https://guides.github.com/activities/citable-code/
