### 2021-07-14 #3159 @tramarobin
@cMadan, great !

> Make a tagged release of your software (that now includes the revisions related to the JOSS review), and list the version tag of the archived version here.

[fctSnPM - v2.0 - JOSS](https://github.com/tramarobin/fctSnPM/releases/tag/v2.0)

> Archive the reviewed software in Zenodo

The software was archived with the same title and the authors as the paper.
The DOI is : 10.5281/zenodo.5101107 and the archive can be found [here](https://zenodo.org/record/5101107#.YO6c4-gzaUk)
