### 2021-07-14 3db08bc
Update paper.bib
Nichols ref modification
