+++ Article@JOSS/paper.bib
@@ -24,7 +24,7 @@
   author  = {{Nichols}, TE. and {Holmes}, AP.}, 
   title   = {{Nonparametric permutation tests for functional neuroimaging: a primer with examples}},
   journal = {Human Brain Mapping},
-  doi = {10.1016/j.jbiomech.2010.03.008},
+  doi = {10.1002/hbm.1058},
   year    = 2002,
   volume = 15,
   pages   = {1-25},
